Extract native/java_util.rs from native/common.rs
