+++ crates/duke-interpreter/src/native/java_util.rs
@@ -1,0 +1,5022 @@
+/// Native: `ZipFile.<init>(String)` — open and index a ZIP/JAR archive.
+pub(crate) fn native_zip_file_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let path_ref = extract_ref_arg(args, 1)?;
+    let path_str = heap
+        .get(path_ref)?
+        .string_value
+        .as_deref()
+        .ok_or(Error::NullPointerException)?
+        .to_string();
+    let fd = zip_open(std::path::Path::new(&path_str))?;
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields[0] = Slot::Int(fd);
+    Ok(None)
+}
+/// Native: `JarFile.<init>(File)` — open and index a JAR archive from a File object.
+pub(crate) fn native_jar_file_init_from_file(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let file_ref = extract_ref_arg(args, 1)?;
+    let path = file_path_from_ref(file_ref, heap)?;
+    let fd = zip_open(&path)?;
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields[0] = Slot::Int(fd);
+    Ok(None)
+}
+pub(crate) fn native_jar_file_init_with_mode_and_version(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let forwarded_args = match args {
+        [this_slot, file_slot, ..] => [*this_slot, *file_slot],
+        _ => {
+            return Err(Error::TypeMismatch {
+                expected: "this,file",
+                got: "other",
+            });
+        }
+    };
+    native_jar_file_init_from_file(&forwarded_args, heap, out, control)
+}
+pub(crate) fn native_jar_file_get_manifest(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fd = extract_io_fd(heap, this_ref)?;
+    let Ok(manifest_bytes) = zip_read_entry(fd, "META-INF/MANIFEST.MF") else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    let manifest_ref = allocate_manifest_from_bytes(heap, &manifest_bytes)?;
+    Ok(Some(Slot::Reference(Some(manifest_ref))))
+}
+/// Native: `ZipFile.getEntry(String) -> ZipEntry` — look up an entry by name.
+#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
+pub(crate) fn native_zip_file_get_entry(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let name_ref = extract_ref_arg(args, 1)?;
+    let fd = extract_io_fd(heap, this_ref)?;
+    let entry_name = heap
+        .get(name_ref)?
+        .string_value
+        .as_deref()
+        .ok_or(Error::NullPointerException)?
+        .to_string();
+    let info = zip_get_entry_info(fd, &entry_name)?;
+    let Some(info) = info else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    // Allocate a ZipEntry HeapObject with 6 fields.
+    let name_heap_ref = heap.allocate_string(info.name);
+    let entry_ref = heap.allocate("java/util/zip/ZipEntry".to_string(), 6);
+    let entry_obj = heap.get_mut(entry_ref)?;
+    entry_obj.fields[0] = Slot::Reference(Some(name_heap_ref));
+    entry_obj.fields[1] = Slot::Int(info.compressed_size as i32);
+    entry_obj.fields[2] = Slot::Int((info.compressed_size >> 32) as i32);
+    entry_obj.fields[3] = Slot::Int(info.uncompressed_size as i32);
+    entry_obj.fields[4] = Slot::Int((info.uncompressed_size >> 32) as i32);
+    entry_obj.fields[5] = Slot::Int(i32::from(info.compression_method));
+    Ok(Some(Slot::Reference(Some(entry_ref))))
+}
+/// Native: `ZipFile.getInputStream(ZipEntry) -> InputStream`
+pub(crate) fn native_zip_file_get_input_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let entry_ref = extract_ref_arg(args, 1)?;
+    let fd = extract_io_fd(heap, this_ref)?;
+    // Get entry name from the ZipEntry object.
+    let name_slot_ref = match heap.get(entry_ref)?.fields.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(Error::NullPointerException),
+    };
+    let entry_name = heap
+        .get(name_slot_ref)?
+        .string_value
+        .as_deref()
+        .ok_or(Error::NullPointerException)?
+        .to_string();
+    // Decompress the entry and wrap in a ByteBuffer.
+    let data = zip_read_entry(fd, &entry_name)?;
+    let buf_fd = heap.open_host_byte_buffer(data);
+    let is_ref = heap.allocate("duke/zip/ByteBufferInputStream".to_string(), 1);
+    let is_obj = heap.get_mut(is_ref)?;
+    is_obj.fields[0] = Slot::Int(buf_fd);
+    Ok(Some(Slot::Reference(Some(is_ref))))
+}
+/// Native: `ZipFile.close()`
+pub(crate) fn native_zip_file_close(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fd = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) => *id,
+        _ => return Ok(None),
+    };
+    zip_close(fd);
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `ZipFile.size() -> int`
+#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
+pub(crate) fn native_zip_file_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fd = extract_io_fd(heap, this_ref)?;
+    let count = zip_entry_count(fd)?;
+    Ok(Some(Slot::Int(count as i32)))
+}
+/// Native: `ZipEntry.getName() -> String`
+pub(crate) fn native_zip_entry_get_name(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    Ok(Some(heap.get(this_ref)?.fields[0]))
+}
+/// Native: `ZipEntry.getCompressedSize() -> long`
+pub(crate) fn native_zip_entry_get_compressed_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fields = &heap.get(this_ref)?.fields;
+    let lo = match fields[1] {
+        Slot::Int(v) => v,
+        _ => 0,
+    };
+    let hi = match fields[2] {
+        Slot::Int(v) => v,
+        _ => 0,
+    };
+    let val = (i64::from(hi) << 32) | (i64::from(lo) & 0xFFFF_FFFF);
+    Ok(Some(Slot::Long(val)))
+}
+/// Native: `ZipEntry.getSize() -> long`
+pub(crate) fn native_zip_entry_get_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fields = &heap.get(this_ref)?.fields;
+    let lo = match fields[3] {
+        Slot::Int(v) => v,
+        _ => 0,
+    };
+    let hi = match fields[4] {
+        Slot::Int(v) => v,
+        _ => 0,
+    };
+    let val = (i64::from(hi) << 32) | (i64::from(lo) & 0xFFFF_FFFF);
+    Ok(Some(Slot::Long(val)))
+}
+/// Native: `ZipEntry.getMethod() -> int`
+pub(crate) fn native_zip_entry_get_method(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    Ok(Some(heap.get(this_ref)?.fields[5]))
+}
+/// Native: `List.of(Object...)List` — all args (fixed-arity or varargs) become a new `ArrayList`.
+///
+/// Handles descriptors with 0–6+ fixed args and the varargs `([O)List` form.
+pub(crate) fn native_list_of(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    // Varargs form: single arg that is an Object[] array.
+    let elems: Vec<Slot> = if args.len() == 1 {
+        if let Some(Slot::Reference(Some(arr_ref))) = args.first() {
+            let obj = heap.get(*arr_ref)?;
+            if obj.class_name.starts_with('[') {
+                let fields = obj.fields.clone();
+                let _ = obj;
+                fields
+            } else {
+                args.to_vec()
+            }
+        } else {
+            Vec::new()
+        }
+    } else {
+        args.to_vec()
+    };
+    let list_ref = make_list_from_slots(&elems, heap, out, control)?;
+    Ok(Some(Slot::Reference(Some(list_ref))))
+}
+/// Native: `Set.of(Object...)Set` — all args (fixed-arity or varargs) become a new `HashSet`.
+///
+/// Handles both fixed-arity descriptors (multiple direct element args)
+/// and the single-array varargs form `([O)Set`.
+pub(crate) fn native_set_of_factory(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let elems: Vec<Slot> = if args.len() == 1 {
+        if let Some(Slot::Reference(Some(arr_ref))) = args.first() {
+            let obj = heap.get(*arr_ref)?;
+            if obj.class_name.starts_with('[') {
+                let fields = obj.fields.clone();
+                let _ = obj;
+                fields
+            } else {
+                args.to_vec()
+            }
+        } else {
+            Vec::new()
+        }
+    } else {
+        args.to_vec()
+    };
+    let set_ref = make_set_from_slots(&elems, heap, out, control)?;
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+/// Native: `Map.of(K,V,...)Map` — pairs of args become entries in a new `HashMap`.
+pub(crate) fn native_map_of(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let map_ref = heap.allocate("java/util/HashMap".to_string(), 1);
+    native_hashmap_init(&[Slot::Reference(Some(map_ref))], heap, out, control)?;
+    let mut i = 0;
+    while i + 1 < args.len() {
+        let k = args[i];
+        let v = args[i + 1];
+        native_hashmap_put(&[Slot::Reference(Some(map_ref)), k, v], heap, out, control)?;
+        i += 2;
+    }
+    Ok(Some(Slot::Reference(Some(map_ref))))
+}
+/// Native: `Optional.empty()Optional` — returns an Optional with no value.
+pub(crate) fn native_optional_empty(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let r = heap.allocate("java/util/Optional".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = Slot::Reference(None);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Optional.of(T)Optional` — wraps value; throws NPE if null.
+pub(crate) fn native_optional_of(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let val = extract_slot_arg(args, 0);
+    if matches!(val, Slot::Reference(None)) {
+        return Err(Error::NullPointerException);
+    }
+    let r = heap.allocate("java/util/Optional".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = val;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Optional.ofNullable(T)Optional` — wraps value or empty if null.
+pub(crate) fn native_optional_of_nullable(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let val = extract_slot_arg(args, 0);
+    let r = heap.allocate("java/util/Optional".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = val;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Optional.get()T` — returns value or throws `NoSuchElementException`.
+pub(crate) fn native_optional_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let val = extract_first_field_arg(heap, this_ref)?;
+    if matches!(val, Slot::Reference(None)) {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    Ok(Some(val))
+}
+/// Native: `Optional.isPresent()Z` — true if a value is present.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_optional_is_present(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let present = !matches!(
+        heap.get(this_ref)?.fields.first(),
+        Some(Slot::Reference(None)) | None
+    );
+    Ok(Some(Slot::Int(i32::from(present))))
+}
+/// Native: `Optional.isEmpty()Z` — true if no value is present (Java 11+).
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_optional_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let empty = matches!(
+        heap.get(this_ref)?.fields.first(),
+        Some(Slot::Reference(None)) | None
+    );
+    Ok(Some(Slot::Int(i32::from(empty))))
+}
+/// Native: `Optional.orElse(T)T` — returns value if present, else the argument.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_optional_or_else(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let val = extract_first_field_arg(heap, this_ref)?;
+    let result = if matches!(val, Slot::Reference(None)) {
+        extract_slot_arg(args, 1)
+    } else {
+        val
+    };
+    Ok(Some(result))
+}
+/// Native: `Optional.orElseThrow()T` — returns value or throws `NoSuchElementException`.
+pub(crate) fn native_optional_or_else_throw(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_optional_get(args, heap, out, control)
+}
+/// Native: `ArrayList.addAll(Collection)Z` — appends all elements from a compatible collection.
+pub(crate) fn native_arraylist_add_all(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let src_ref = extract_ref_arg(args, 1)?;
+    let src_size = match heap.get(src_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => return Ok(Some(Slot::Int(0))),
+    };
+    let elems: Vec<Slot> = heap.get(src_ref)?.fields[1..=src_size].to_vec();
+    let modified = !elems.is_empty();
+    for elem in elems {
+        native_arraylist_add(&[Slot::Reference(Some(this_ref)), elem], heap, out, control)?;
+    }
+    Ok(Some(Slot::Int(i32::from(modified))))
+}
+/// Native: `HashMap.putAll(Map)V` — copies all entries from the source `HashMap`.
+pub(crate) fn native_hashmap_put_all(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let src_ref = extract_ref_arg(args, 1)?;
+    let fields = heap.get(src_ref)?.fields.clone();
+    // fields[0] = size, fields[1..] = k0, v0, k1, v1, ...
+    let mut i = 1;
+    while i + 1 < fields.len() {
+        let k = fields[i];
+        let v = fields[i + 1];
+        native_hashmap_put(&[Slot::Reference(Some(this_ref)), k, v], heap, out, control)?;
+        i += 2;
+    }
+    Ok(None)
+}
+/// Native: `HashMap.computeIfAbsent(K, Function)V` — returns existing value or computes and stores it.
+pub(crate) fn native_hashmap_compute_if_absent(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    // Check if key already present.
+    let existing = native_hashmap_get(&[Slot::Reference(Some(this_ref)), key], heap, out, control)?;
+    if let Some(v) = existing
+        && !matches!(v, Slot::Reference(None))
+    {
+        return Ok(Some(v));
+    }
+    // Key absent — invoke the mapping function (lambda / SAM).
+    let fn_ref = extract_ref_arg(args, 2)?;
+    let fn_class = heap.get(fn_ref)?.class_name.clone();
+    let computed = ops.invoke(
+        heap,
+        out,
+        &fn_class,
+        "apply",
+        "(Ljava/lang/Object;)Ljava/lang/Object;",
+        vec![Slot::Reference(Some(fn_ref)), key],
+    )?;
+    if let Some(value) = computed
+        && !matches!(value, Slot::Reference(None))
+    {
+        native_hashmap_put(
+            &[Slot::Reference(Some(this_ref)), key, value],
+            heap,
+            out,
+            control,
+        )?;
+        return Ok(Some(value));
+    }
+    Ok(Some(Slot::Reference(None)))
+}
+/// Native: `LinkedList.<init>()V` — same initialisation as `ArrayList`.
+pub(crate) fn native_linked_list_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_init(args, heap, out, control)
+}
+/// Native: `LinkedList.<init>(Collection)V` — copies all elements from source collection.
+pub(crate) fn native_linked_list_init_collection(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let src_ref = extract_ref_arg(args, 1)?;
+    let src_size = match heap.get(src_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let src_elems: Vec<Slot> = heap.get(src_ref)?.fields[1..=src_size].to_vec();
+    let n = i32::try_from(src_elems.len()).unwrap_or(0);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(n);
+    heap.get_mut(this_ref)?.fields.extend(src_elems);
+    Ok(None)
+}
+/// Native: `LinkedList.size()I`
+pub(crate) fn native_linked_list_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_size(args, heap, out, control)
+}
+/// Native: `LinkedList.add(Object)Z` — appends to tail.
+pub(crate) fn native_linked_list_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)
+}
+/// Native: `LinkedList.get(I)Object`
+pub(crate) fn native_linked_list_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_get(args, heap, out, control)
+}
+/// Native: `LinkedList.addFirst(Object)V` — inserts at index 0.
+/// Field layout: `fields[0]`=Int(size), `fields[1..size]`=elements.
+pub(crate) fn native_linked_list_add_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elem = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let obj = heap.get_mut(this_ref)?;
+    // Shift existing elements right by one.
+    obj.fields.insert(1, elem);
+    obj.fields[0] = Slot::Int(i32::try_from(size + 1).unwrap_or(i32::MAX));
+    Ok(None)
+}
+/// Native: `LinkedList.addLast(Object)V` — appends to tail (same as add).
+pub(crate) fn native_linked_list_add_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)?;
+    Ok(None)
+}
+/// Native: `LinkedList.peekFirst()Object` — returns head without removal, or null if empty.
+pub(crate) fn native_linked_list_peek_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => *n,
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    Ok(Some(heap.get(this_ref)?.fields[1]))
+}
+/// Native: `LinkedList.peekLast()Object` — returns tail without removal, or null if empty.
+pub(crate) fn native_linked_list_peek_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    Ok(Some(heap.get(this_ref)?.fields[size]))
+}
+/// Native: `LinkedList.removeFirst()Object` — removes and returns head.
+pub(crate) fn native_linked_list_remove_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    let obj = heap.get_mut(this_ref)?;
+    let elem = obj.fields.remove(1);
+    obj.fields[0] = Slot::Int(i32::try_from(size - 1).unwrap_or(0));
+    Ok(Some(elem))
+}
+/// Native: `LinkedList.removeLast()Object` — removes and returns tail.
+pub(crate) fn native_linked_list_remove_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    let obj = heap.get_mut(this_ref)?;
+    let elem = obj.fields.remove(size);
+    obj.fields[0] = Slot::Int(i32::try_from(size - 1).unwrap_or(0));
+    Ok(Some(elem))
+}
+/// Native: `LinkedList.poll()Object` — removes and returns head, or null if empty.
+pub(crate) fn native_linked_list_poll(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    let obj = heap.get_mut(this_ref)?;
+    let elem = obj.fields.remove(1);
+    obj.fields[0] = Slot::Int(i32::try_from(size - 1).unwrap_or(0));
+    Ok(Some(elem))
+}
+/// Native: `LinkedList.offer(Object)Z` — appends to tail, returns true.
+pub(crate) fn native_linked_list_offer(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)
+}
+/// Native: `LinkedList.isEmpty()Z`
+pub(crate) fn native_linked_list_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_is_empty(args, heap, out, control)
+}
+/// Native: `LinkedList.iterator()Iterator` — returns an ArrayList-compatible iterator.
+pub(crate) fn native_linked_list_iterator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_iterator(args, heap, out, control)
+}
+/// Native: `HashMap.forEach(BiConsumer)V` — iterates key-value pairs, invoking `accept(k, v)`.
+pub(crate) fn native_hashmap_for_each(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let consumer_ref = extract_ref_arg(args, 1)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    // Snapshot key-val pairs (fields[1,2], fields[3,4], ...)
+    let pairs: Vec<(Slot, Slot)> = (0..size)
+        .map(|i| {
+            let key = heap.get(this_ref).map_or(Slot::Reference(None), |o| o.fields[1 + i * 2]);
+            let val = heap.get(this_ref).map_or(Slot::Reference(None), |o| o.fields[2 + i * 2]);
+            (key, val)
+        })
+        .collect();
+    let consumer_class = heap.get(consumer_ref)?.class_name.clone();
+    for (key, val) in pairs {
+        ops.invoke(
+            heap,
+            out,
+            &consumer_class,
+            "accept",
+            "(Ljava/lang/Object;Ljava/lang/Object;)V",
+            vec![Slot::Reference(Some(consumer_ref)), key, val],
+        )?;
+    }
+    let _ = control;
+    Ok(None)
+}
+/// Native: `HashMap.replaceAll(BiFunction<K,V,V>) -> void`
+/// Replaces each value with the result of applying the function to (key, value).
+pub(crate) fn native_hashmap_replace_all(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fn_ref = extract_ref_arg(args, 1)?;
+    let fn_class = heap.get(fn_ref)?.class_name.clone();
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    // Snapshot keys (values will be mutated in place).
+    let keys: Vec<Slot> = (0..size)
+        .map(|i| {
+            heap.get(this_ref).map_or(Slot::Reference(None), |o| o.fields[1 + i * 2])
+        })
+        .collect();
+    for (i, key) in keys.iter().enumerate() {
+        let old_val = heap.get(this_ref).map_or(Slot::Reference(None), |o| o.fields[2 + i * 2]);
+        let new_val = ops.invoke(
+            heap,
+            out,
+            &fn_class,
+            "apply",
+            "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
+            vec![Slot::Reference(Some(fn_ref)), *key, old_val],
+        )?;
+        if let Some(v) = new_val {
+            heap.get_mut(this_ref)?.fields[2 + i * 2] = v;
+        }
+    }
+    let _ = control;
+    Ok(None)
+}
+/// Native: `TreeMap.<init>()V`
+pub(crate) fn native_treemap_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `TreeMap.put(K,V)V` — inserts in sorted key order.
+pub(crate) fn native_treemap_put(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let val = extract_slot_arg(args, 2);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    // Check for existing key — update in place.
+    for i in 0..size {
+        let existing_key = heap.get(this_ref)?.fields[1 + i * 2];
+        if treemap_keys_equal(existing_key, key, heap) {
+            heap.get_mut(this_ref)?.fields[2 + i * 2] = val;
+            return Ok(Some(Slot::Reference(None)));
+        }
+    }
+    // Find sorted insert position.
+    let insert_pos = {
+        let mut pos = size;
+        for i in 0..size {
+            let ek = heap.get(this_ref)?.fields[1 + i * 2];
+            if compare_treemap_keys(key, ek, heap) == std::cmp::Ordering::Less {
+                pos = i;
+                break;
+            }
+        }
+        pos
+    };
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields.insert(1 + insert_pos * 2, val);
+    obj.fields.insert(1 + insert_pos * 2, key);
+    obj.fields[0] = Slot::Int(i32::try_from(size + 1).unwrap_or(i32::MAX));
+    Ok(Some(Slot::Reference(None)))
+}
+/// Native: `TreeMap.get(K)V`
+pub(crate) fn native_treemap_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    for i in 0..size {
+        let ek = heap.get(this_ref)?.fields[1 + i * 2];
+        if treemap_keys_equal(ek, key, heap) {
+            let v = heap.get(this_ref)?.fields[2 + i * 2];
+            return Ok(Some(v));
+        }
+    }
+    Ok(Some(Slot::Reference(None)))
+}
+/// Native: `TreeMap.containsKey(K)Z`
+pub(crate) fn native_treemap_contains_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let found = !matches!(
+        native_treemap_get(args, heap, out, control)?,
+        Some(Slot::Reference(None)) | None
+    );
+    Ok(Some(Slot::Int(i32::from(found))))
+}
+/// Native: `TreeMap.size()I`
+pub(crate) fn native_treemap_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    Ok(Some(match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => Slot::Int(*n),
+        _ => Slot::Int(0),
+    }))
+}
+/// Native: `TreeMap.firstKey()K` — returns the smallest key (index 0 in sorted list).
+pub(crate) fn native_treemap_first_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => *n,
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    Ok(Some(heap.get(this_ref)?.fields[1]))
+}
+/// Native: `TreeMap.lastKey()K` — returns the largest key.
+pub(crate) fn native_treemap_last_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    Ok(Some(heap.get(this_ref)?.fields[1 + (size - 1) * 2]))
+}
+/// Native: `TreeMap.remove(K)V`
+pub(crate) fn native_treemap_remove(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    for i in 0..size {
+        let ek = heap.get(this_ref)?.fields[1 + i * 2];
+        if treemap_keys_equal(ek, key, heap) {
+            let obj = heap.get_mut(this_ref)?;
+            let v = obj.fields.remove(2 + i * 2);
+            obj.fields.remove(1 + i * 2);
+            obj.fields[0] = Slot::Int(i32::try_from(size - 1).unwrap_or(0));
+            return Ok(Some(v));
+        }
+    }
+    Ok(Some(Slot::Reference(None)))
+}
+/// Native: `TreeMap.isEmpty()Z`
+pub(crate) fn native_treemap_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    Ok(Some(Slot::Int(match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(0)) | None => 1,
+        _ => 0,
+    })))
+}
+/// Native: `TreeMap.headMap(toKey)SortedMap` — returns a new `TreeMap` with keys strictly less than toKey.
+pub(crate) fn native_treemap_head_map(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let to_key = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let result = heap.allocate("java/util/TreeMap".to_string(), 1);
+    heap.get_mut(result)?.fields[0] = Slot::Int(0);
+    let mut count = 0usize;
+    for i in 0..size {
+        let k = heap.get(this_ref)?.fields[1 + i * 2];
+        let v = heap.get(this_ref)?.fields[2 + i * 2];
+        if compare_treemap_keys(k, to_key, heap) == std::cmp::Ordering::Less {
+            heap.get_mut(result)?.fields.push(k);
+            heap.get_mut(result)?.fields.push(v);
+            count += 1;
+        }
+    }
+    heap.get_mut(result)?.fields[0] = Slot::Int(i32::try_from(count).unwrap_or(0));
+    Ok(Some(Slot::Reference(Some(result))))
+}
+/// Native: `TreeMap.tailMap(fromKey)SortedMap` — returns a new `TreeMap` with keys >= fromKey.
+pub(crate) fn native_treemap_tail_map(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let from_key = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let result = heap.allocate("java/util/TreeMap".to_string(), 1);
+    heap.get_mut(result)?.fields[0] = Slot::Int(0);
+    let mut count = 0usize;
+    for i in 0..size {
+        let k = heap.get(this_ref)?.fields[1 + i * 2];
+        let v = heap.get(this_ref)?.fields[2 + i * 2];
+        if compare_treemap_keys(k, from_key, heap) != std::cmp::Ordering::Less {
+            heap.get_mut(result)?.fields.push(k);
+            heap.get_mut(result)?.fields.push(v);
+            count += 1;
+        }
+    }
+    heap.get_mut(result)?.fields[0] = Slot::Int(i32::try_from(count).unwrap_or(0));
+    Ok(Some(Slot::Reference(Some(result))))
+}
+/// Native: `Stack.<init>()V`
+pub(crate) fn native_stack_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_init(args, heap, out, control)
+}
+/// Native: `Stack.push(E)E` — appends to tail, returns the element.
+pub(crate) fn native_stack_push(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let elem = extract_slot_arg(args, 1);
+    native_arraylist_add(args, heap, out, control)?;
+    Ok(Some(elem))
+}
+/// Native: `Stack.pop()E` — removes and returns the top element.
+pub(crate) fn native_stack_pop(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_linked_list_remove_last(args, heap, out, control)
+}
+/// Native: `Stack.peek()E` — returns the top element without removal.
+pub(crate) fn native_stack_peek(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_linked_list_peek_last(args, heap, out, control)
+}
+/// Native: `Stack.empty()Z` — returns true if the stack is empty.
+pub(crate) fn native_stack_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_is_empty(args, heap, out, control)
+}
+/// Native: `Stack.size()I`
+pub(crate) fn native_stack_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_size(args, heap, out, control)
+}
+/// Native: `TreeSet.<init>()V`
+pub(crate) fn native_treeset_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+pub(crate) fn native_treeset_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elem = extract_slot_arg(args, 1);
+    let elem_key = treeset_slot_sort_key(elem, heap);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    // Check for duplicate.
+    for i in 0..size {
+        let ex = heap.get(this_ref)?.fields[1 + i];
+        let ex_key = treeset_slot_sort_key(ex, heap);
+        if ex_key == elem_key {
+            return Ok(Some(Slot::Int(0))); // false — no change
+        }
+    }
+    // Find sorted insert position.
+    let insert_pos = {
+        let mut pos = size;
+        for i in 0..size {
+            let ex = heap.get(this_ref)?.fields[1 + i];
+            let ex_key = treeset_slot_sort_key(ex, heap);
+            if let (Some(ek), Some(exk)) = (&elem_key, &ex_key)
+                && ek.less_than(exk)
+            {
+                pos = i;
+                break;
+            }
+        }
+        pos
+    };
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields.insert(1 + insert_pos, elem);
+    obj.fields[0] = Slot::Int(i32::try_from(size + 1).unwrap_or(i32::MAX));
+    Ok(Some(Slot::Int(1))) // true — element added
+}
+/// Native: `TreeSet.contains(E)Z`
+pub(crate) fn native_treeset_contains(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elem = extract_slot_arg(args, 1);
+    let elem_str = match &elem {
+        Slot::Reference(Some(r)) => heap.get(*r)?.string_value.clone(),
+        Slot::Int(v) => Some(v.to_string()),
+        _ => None,
+    };
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    for i in 0..size {
+        let ex = heap.get(this_ref)?.fields[1 + i];
+        let ex_str = match &ex {
+            Slot::Reference(Some(r)) => heap.get(*r).ok().and_then(|o| o.string_value.clone()),
+            Slot::Int(v) => Some(v.to_string()),
+            _ => None,
+        };
+        if ex_str == elem_str {
+            return Ok(Some(Slot::Int(1)));
+        }
+    }
+    Ok(Some(Slot::Int(0)))
+}
+/// Native: `TreeSet.size()I`
+pub(crate) fn native_treeset_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    Ok(Some(match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => Slot::Int(*n),
+        _ => Slot::Int(0),
+    }))
+}
+/// Native: `TreeSet.first()E` — returns smallest element.
+pub(crate) fn native_treeset_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    match heap.get(this_ref)?.fields.first().copied() {
+        Some(Slot::Int(0)) | None => Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        }),
+        _ => Ok(Some(heap.get(this_ref)?.fields[1])),
+    }
+}
+/// Native: `TreeSet.last()E` — returns largest element.
+pub(crate) fn native_treeset_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    Ok(Some(heap.get(this_ref)?.fields[size]))
+}
+/// Native: `TreeSet.isEmpty()Z`
+pub(crate) fn native_treeset_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    Ok(Some(Slot::Int(match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(0)) | None => 1,
+        _ => 0,
+    })))
+}
+/// Native: `TreeSet.iterator()Iterator` — returns an ArrayList-compatible iterator over sorted elements.
+pub(crate) fn native_treeset_iterator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_iterator(args, heap, out, control)
+}
+/// Native: `Collections.min(Collection)T` — returns minimum element via `compareTo`.
+pub(crate) fn native_collections_min(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let coll_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(coll_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    let Slot::Reference(Some(mut min_ref)) = heap.get(coll_ref)?.fields[1] else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    for i in 2..=size {
+        let Slot::Reference(Some(candidate)) = heap.get(coll_ref)?.fields[i] else {
+            continue;
+        };
+        let class_name = heap.get(candidate)?.class_name.clone();
+        let cmp = ops.invoke(
+            heap,
+            out,
+            &class_name,
+            "compareTo",
+            "(Ljava/lang/Object;)I",
+            vec![
+                Slot::Reference(Some(candidate)),
+                Slot::Reference(Some(min_ref)),
+            ],
+        )?;
+        let _ = control;
+        if matches!(cmp, Some(Slot::Int(n)) if n < 0) {
+            min_ref = candidate;
+        }
+    }
+    Ok(Some(Slot::Reference(Some(min_ref))))
+}
+/// Native: `Collections.max(Collection)T` — returns maximum element via `compareTo`.
+pub(crate) fn native_collections_max(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let coll_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(coll_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    let Slot::Reference(Some(mut max_ref)) = heap.get(coll_ref)?.fields[1] else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    for i in 2..=size {
+        let Slot::Reference(Some(candidate)) = heap.get(coll_ref)?.fields[i] else {
+            continue;
+        };
+        let class_name = heap.get(candidate)?.class_name.clone();
+        let cmp = ops.invoke(
+            heap,
+            out,
+            &class_name,
+            "compareTo",
+            "(Ljava/lang/Object;)I",
+            vec![
+                Slot::Reference(Some(candidate)),
+                Slot::Reference(Some(max_ref)),
+            ],
+        )?;
+        let _ = control;
+        if matches!(cmp, Some(Slot::Int(n)) if n > 0) {
+            max_ref = candidate;
+        }
+    }
+    Ok(Some(Slot::Reference(Some(max_ref))))
+}
+/// Native: `Collections.shuffle(List)V` — no-op (deterministic test environments).
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_collections_shuffle(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    Ok(None)
+}
+/// Native: `Collections.shuffle(List, Random)V` — shuffle with provided RNG (no-op for correctness since test only checks sum).
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_collections_shuffle_random(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    Ok(None)
+}
+/// Native: `Collections.fill(List, Object)V` — set every element to value.
+pub(crate) fn native_collections_fill(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let value = extract_slot_arg(args, 1);
+    let size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    for i in 1..=size {
+        heap.get_mut(list_ref)?.fields[i] = value;
+    }
+    Ok(None)
+}
+/// Native: `ArrayList.stream()` — wrap `ArrayList` elements into a `Stream`.
+pub(crate) fn native_arraylist_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(n)) => *n,
+        _ => 0,
+    };
+    let elems: Vec<Slot> =
+        heap.get(list_ref)?.fields[1..=usize::try_from(size).unwrap_or(0)].to_vec();
+    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
+    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(size);
+    heap.get_mut(stream_ref)?.fields.extend(elems);
+    Ok(Some(Slot::Reference(Some(stream_ref))))
+}
+/// Native: `ArrayDeque.<init>()V` — same layout as `ArrayList`.
+pub(crate) fn native_arraydeque_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_init(args, heap, out, control)
+}
+/// Native: `ArrayDeque.push(Object)V` — push to front (stack: LIFO).
+pub(crate) fn native_arraydeque_push(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elem = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    heap.get_mut(this_ref)?.fields.insert(1, elem);
+    let new_size = i32::try_from(size + 1).unwrap_or(0);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(new_size);
+    Ok(None)
+}
+/// Native: `ArrayDeque.pop()Object` — pop from front (stack: LIFO).
+pub(crate) fn native_arraydeque_pop(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Err(Error::JavaException {
+            class_name: "java/util/NoSuchElementException".to_string(),
+        });
+    }
+    let elem = heap.get_mut(this_ref)?.fields.remove(1);
+    let new_size = i32::try_from(size - 1).unwrap_or(0);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(new_size);
+    Ok(Some(elem))
+}
+/// Native: `ArrayDeque.offer(Object)Z` — enqueue at back (queue: FIFO).
+pub(crate) fn native_arraydeque_offer(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)?;
+    Ok(Some(Slot::Int(1))) // always succeeds
+}
+/// Native: `ArrayDeque.add(Object)Z` — same as offer (appends to back).
+pub(crate) fn native_arraydeque_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)?;
+    Ok(Some(Slot::Int(1)))
+}
+/// Native: `ArrayDeque.poll()Object` — dequeue from front (queue: FIFO); null if empty.
+pub(crate) fn native_arraydeque_poll(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    let elem = heap.get_mut(this_ref)?.fields.remove(1);
+    let new_size = i32::try_from(size - 1).unwrap_or(0);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(new_size);
+    Ok(Some(elem))
+}
+/// Native: `ArrayDeque.peek()Object` — peek at front; null if empty.
+pub(crate) fn native_arraydeque_peek(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    Ok(Some(heap.get(this_ref)?.fields[1]))
+}
+/// Native: `ArrayDeque.size()I`
+pub(crate) fn native_arraydeque_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_size(args, heap, out, control)
+}
+/// Native: `ArrayDeque.isEmpty()Z`
+pub(crate) fn native_arraydeque_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_is_empty(args, heap, out, control)
+}
+/// Native: `PriorityQueue.<init>()V` — same init as `ArrayList`.
+pub(crate) fn native_priorityqueue_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_init(args, heap, out, control)
+}
+/// Native: `PriorityQueue.offer(Object)Z` — inserts in heap order via `compareTo`.
+pub(crate) fn native_priorityqueue_offer(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elem = extract_slot_arg(args, 1);
+    // Append, then sift up.
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    heap.get_mut(this_ref)?.fields.push(elem);
+    let new_size = size + 1;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(i32::try_from(new_size).unwrap_or(0));
+    // Sift up from last position (1-indexed in fields).
+    let mut i = new_size; // fields index of newly added element
+    while i > 1 {
+        // fields are 1-indexed: child at fields[i], parent at fields[i/2]
+        let parent_idx = i / 2;
+        let child_slot = heap.get(this_ref)?.fields[i];
+        let parent_slot = heap.get(this_ref)?.fields[parent_idx];
+        let (Slot::Reference(Some(child_ref)), Slot::Reference(Some(parent_ref))) =
+            (child_slot, parent_slot)
+        else {
+            break;
+        };
+        let class_child = heap.get(child_ref)?.class_name.clone();
+        let cmp = ops.invoke(
+            heap,
+            out,
+            &class_child,
+            "compareTo",
+            "(Ljava/lang/Object;)I",
+            vec![
+                Slot::Reference(Some(child_ref)),
+                Slot::Reference(Some(parent_ref)),
+            ],
+        )?;
+        if matches!(cmp, Some(Slot::Int(n)) if n < 0) {
+            // child < parent: swap
+            heap.get_mut(this_ref)?.fields.swap(i, parent_idx);
+            i = parent_idx;
+        } else {
+            break;
+        }
+    }
+    Ok(Some(Slot::Int(1)))
+}
+/// Native: `PriorityQueue.add(Object)Z` — same as offer.
+pub(crate) fn native_priorityqueue_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    native_priorityqueue_offer(args, heap, out, control, ops)
+}
+/// Native: `PriorityQueue.peek()Object` — returns minimum element without removing.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_priorityqueue_peek(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    Ok(Some(heap.get(this_ref)?.fields[1]))
+}
+/// Native: `PriorityQueue.poll()Object` — removes and returns minimum; sifts down.
+pub(crate) fn native_priorityqueue_poll(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    let min = heap.get(this_ref)?.fields[1];
+    if size == 1 {
+        heap.get_mut(this_ref)?.fields.pop();
+        heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+        return Ok(Some(min));
+    }
+    // Move last element to root, sift down.
+    let last = heap.get(this_ref)?.fields[size];
+    heap.get_mut(this_ref)?.fields[1] = last;
+    heap.get_mut(this_ref)?.fields.pop();
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(i32::try_from(size - 1).unwrap_or(0));
+    let new_size = size - 1;
+    let mut i = 1usize;
+    loop {
+        let left = 2 * i;
+        let right = 2 * i + 1;
+        let mut smallest = i;
+        if left <= new_size {
+            let (cur_slot, left_slot) = (
+                heap.get(this_ref)?.fields[smallest],
+                heap.get(this_ref)?.fields[left],
+            );
+            let (Slot::Reference(Some(cur_ref)), Slot::Reference(Some(left_ref))) =
+                (cur_slot, left_slot)
+            else {
+                break;
+            };
+            let class_left = heap.get(left_ref)?.class_name.clone();
+            let cmp = ops.invoke(
+                heap,
+                out,
+                &class_left,
+                "compareTo",
+                "(Ljava/lang/Object;)I",
+                vec![
+                    Slot::Reference(Some(left_ref)),
+                    Slot::Reference(Some(cur_ref)),
+                ],
+            )?;
+            if matches!(cmp, Some(Slot::Int(n)) if n < 0) {
+                smallest = left;
+            }
+        }
+        if right <= new_size {
+            let (small_slot, right_slot) = (
+                heap.get(this_ref)?.fields[smallest],
+                heap.get(this_ref)?.fields[right],
+            );
+            let (Slot::Reference(Some(small_ref)), Slot::Reference(Some(right_ref))) =
+                (small_slot, right_slot)
+            else {
+                break;
+            };
+            let class_right = heap.get(right_ref)?.class_name.clone();
+            let cmp = ops.invoke(
+                heap,
+                out,
+                &class_right,
+                "compareTo",
+                "(Ljava/lang/Object;)I",
+                vec![
+                    Slot::Reference(Some(right_ref)),
+                    Slot::Reference(Some(small_ref)),
+                ],
+            )?;
+            if matches!(cmp, Some(Slot::Int(n)) if n < 0) {
+                smallest = right;
+            }
+        }
+        if smallest == i {
+            break;
+        }
+        heap.get_mut(this_ref)?.fields.swap(i, smallest);
+        i = smallest;
+    }
+    Ok(Some(min))
+}
+/// Native: `PriorityQueue.size()I`
+pub(crate) fn native_priorityqueue_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_size(args, heap, out, control)
+}
+/// Native: `PriorityQueue.isEmpty()Z`
+pub(crate) fn native_priorityqueue_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_is_empty(args, heap, out, control)
+}
+/// Native: `Comparator.naturalOrder()Comparator` — returns a singleton synthetic comparator.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_comparator_natural_order(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let r = heap.allocate("duke/util/NaturalOrderComparator".to_string(), 0);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Comparator.reverseOrder()Comparator` — returns a singleton reverse comparator.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_comparator_reverse_order(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let r = heap.allocate("duke/util/ReverseOrderComparator".to_string(), 0);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Comparator.comparingInt(ToIntFunction)Comparator` — wraps key extractor.
+/// Creates a `duke/util/ComparingIntComparator` with `fields[0] = fn_ref`.
+pub(crate) fn native_comparator_comparing_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let fn_ref = extract_ref_arg(args, 0)?;
+    let r = heap.allocate("duke/util/ComparingIntComparator".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = Slot::Reference(Some(fn_ref));
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Collections.sort(List, Comparator)V` — 2-arg sort with explicit comparator.
+pub(crate) fn native_collections_sort_with_comparator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let comparator = extract_slot_arg(args, 1);
+    let class_name = heap.get(list_ref)?.class_name.clone();
+    ops.invoke(
+        heap,
+        output,
+        &class_name,
+        "sort",
+        "(Ljava/util/Comparator;)V",
+        vec![Slot::Reference(Some(list_ref)), comparator],
+    )?;
+    Ok(None)
+}
+pub(crate) fn native_manifest_get_main_attributes(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let manifest_ref = extract_ref_arg(args, 0)?;
+    let Some(Slot::Reference(Some(raw_ref))) = heap.get(manifest_ref)?.fields.first().copied()
+    else {
+        return Err(Error::NullPointerException);
+    };
+    let attributes_ref = heap.allocate("java/util/jar/Attributes".to_string(), 1);
+    heap.get_mut(attributes_ref)?.fields[0] = Slot::Reference(Some(raw_ref));
+    Ok(Some(Slot::Reference(Some(attributes_ref))))
+}
+pub(crate) fn native_attributes_get_value(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let attributes_ref = extract_ref_arg(args, 0)?;
+    let key_ref = extract_ref_arg(args, 1)?;
+    let Some(Slot::Reference(Some(raw_ref))) = heap.get(attributes_ref)?.fields.first().copied()
+    else {
+        return Err(Error::NullPointerException);
+    };
+    let manifest_text = string_value_from_ref(heap, raw_ref)?;
+    let key = string_value_from_ref(heap, key_ref)?;
+    let result = manifest_attribute_value(manifest_text.as_bytes(), &key)
+        .map_or(Slot::Reference(None), |value| {
+            Slot::Reference(Some(heap.allocate_string(value)))
+        });
+    Ok(Some(result))
+}
+/// Native: `Objects.isNull(Object)Z` — returns 1 if argument is null.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_objects_is_null(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let is_null = matches!(args.first(), Some(Slot::Reference(None)) | None);
+    Ok(Some(Slot::Int(i32::from(is_null))))
+}
+/// Native: `Objects.nonNull(Object)Z` — returns 1 if argument is not null.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_objects_non_null(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let is_null = matches!(args.first(), Some(Slot::Reference(None)) | None);
+    Ok(Some(Slot::Int(i32::from(!is_null))))
+}
+/// Native: `Objects.requireNonNull(Object)Object` — throws NPE if null, else returns arg.
+pub(crate) fn native_objects_require_non_null(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    match args.first() {
+        Some(Slot::Reference(None)) | None => Err(Error::NullPointerException),
+        Some(s) => Ok(Some(*s)),
+    }
+}
+/// Native: `Objects.requireNonNull(Object, String)Object` — throws NPE with message if null.
+pub(crate) fn native_objects_require_non_null_msg(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    match args.first() {
+        Some(Slot::Reference(None)) | None => Err(Error::NullPointerException),
+        Some(s) => Ok(Some(*s)),
+    }
+}
+/// Native: `Objects.equals(Object, Object)Z` — null-safe equality check.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_objects_equals(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_slot_arg(args, 0);
+    let b = extract_slot_arg(args, 1);
+    let equal = slots_equal(&a, &b, heap);
+    Ok(Some(Slot::Int(i32::from(equal))))
+}
+/// Native: `Objects.toString(Object)` — returns `"null"` if null, else `string_value` or class name.
+pub(crate) fn native_objects_tostring(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let s = match args.first() {
+        Some(Slot::Reference(None)) | None => heap.allocate_string("null".to_string()),
+        Some(Slot::Reference(Some(r))) => {
+            let obj = heap.get(*r)?;
+            let text = obj
+                .string_value
+                .as_deref()
+                .map_or_else(|| format!("{}@{}", obj.class_name, r), str::to_owned);
+            let _ = obj;
+            heap.allocate_string(text)
+        }
+        Some(Slot::Int(n)) => heap.allocate_string(n.to_string()),
+        Some(Slot::Long(n)) => heap.allocate_string(n.to_string()),
+        Some(other) => heap.allocate_string(format!("{other:?}")),
+    };
+    Ok(Some(Slot::Reference(Some(s))))
+}
+/// Native: `Objects.toString(Object, String)String` — returns nullDefault if null, else toString.
+pub(crate) fn native_objects_tostring_default(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    match args.first() {
+        Some(Slot::Reference(None)) | None => {
+            // null → return the default string (args[1])
+            let default_ref = match args.get(1) {
+                Some(Slot::Reference(Some(r))) => *r,
+                _ => heap.allocate_string("null".to_string()),
+            };
+            Ok(Some(Slot::Reference(Some(default_ref))))
+        }
+        Some(Slot::Reference(Some(r))) => {
+            let obj = heap.get(*r)?;
+            let text = obj
+                .string_value
+                .as_deref()
+                .map_or_else(|| format!("{}@{}", obj.class_name, r), str::to_owned);
+            let _ = obj;
+            let s = heap.allocate_string(text);
+            Ok(Some(Slot::Reference(Some(s))))
+        }
+        Some(Slot::Int(n)) => {
+            let s = heap.allocate_string(n.to_string());
+            Ok(Some(Slot::Reference(Some(s))))
+        }
+        Some(other) => {
+            let s = heap.allocate_string(format!("{other:?}"));
+            Ok(Some(Slot::Reference(Some(s))))
+        }
+    }
+}
+/// Native: `Objects.hashCode(Object)I` — returns 0 for null, else the object's
+/// stable identity hash (relocation-safe; masked non-negative).
+pub(crate) fn native_objects_hashcode(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let hash = match args.first() {
+        Some(Slot::Reference(Some(r))) => heap.identity_hash(*r)? & 0x7FFF_FFFF,
+        _ => 0,
+    };
+    Ok(Some(Slot::Int(hash)))
+}
+/// Native: `Collections.emptyList()List` — returns a new empty `ArrayList`.
+pub(crate) fn native_collections_empty_list(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    // Return an immutable empty UnmodifiableList (same field layout as ArrayList)
+    let r = heap.allocate("java/util/UnmodifiableList".to_string(), 1);
+    native_arraylist_init(&[Slot::Reference(Some(r))], heap, out, control)?;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Collections.emptySet()Set` — returns a new empty `HashSet`.
+pub(crate) fn native_collections_empty_set(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let r = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(r))], heap, out, control)?;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Collections.emptyMap()Map` — returns a new empty `HashMap`.
+pub(crate) fn native_collections_empty_map(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let r = heap.allocate("java/util/HashMap".to_string(), 1);
+    native_hashmap_init(&[Slot::Reference(Some(r))], heap, out, control)?;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Collections.unmodifiableList(List)List` — returns the same list (no-copy; single-threaded).
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_collections_unmodifiable_list(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    // Create an UnmodifiableList backed by the source list's elements.
+    // Mutation methods on this class throw UnsupportedOperationException.
+    let src_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Ok(Some(Slot::Reference(None))),
+    };
+    let (size, elems) = {
+        let src = heap.get(src_ref)?;
+        let size = src.fields.first().copied().unwrap_or(Slot::Int(0));
+        let elems = src.fields[1..].to_vec();
+        (size, elems)
+    };
+    let n_fields = 1 + elems.len();
+    let dst_ref = heap.allocate("java/util/UnmodifiableList".to_string(), n_fields);
+    {
+        let dst = heap.get_mut(dst_ref)?;
+        dst.fields[0] = size;
+        for (i, e) in elems.iter().enumerate() {
+            dst.fields[1 + i] = *e;
+        }
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+/// Native: mutation ops on `UnmodifiableList` throw `UnsupportedOperationException`.
+pub(crate) fn native_unmodifiable_list_mutation(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    Err(Error::JavaException {
+        class_name: "java/lang/UnsupportedOperationException".to_string(),
+    })
+}
+/// Native: `Arrays.stream(int[])IntStream` — wraps an int array into an `IntStream`.
+pub(crate) fn native_arrays_stream_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    // int[] layout: fields = Int elements directly (no length header); arraylength = fields.len()
+    let arr_obj = heap.get(arr_ref)?;
+    let values: Vec<i32> = arr_obj
+        .fields
+        .iter()
+        .filter_map(|s| if let Slot::Int(n) = s { Some(*n) } else { None })
+        .collect();
+    Ok(Some(Slot::Reference(Some(make_int_stream(heap, values)))))
+}
+/// Native: `Arrays.stream(int[], int, int)IntStream` — wraps a subrange as an `IntStream`.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_arrays_stream_int_range(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let from = match args.get(1) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let to = match args.get(2) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let arr_obj = heap.get(arr_ref)?;
+    let values: Vec<i32> = arr_obj
+        .fields
+        .get(from..to.min(arr_obj.fields.len()))
+        .unwrap_or(&[])
+        .iter()
+        .filter_map(|s| if let Slot::Int(n) = s { Some(*n) } else { None })
+        .collect();
+    Ok(Some(Slot::Reference(Some(make_int_stream(heap, values)))))
+}
+/// Native: `Arrays.stream(Object[])Stream` — wraps a reference array as an eager `Stream`.
+pub(crate) fn native_arrays_stream_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let elems: Vec<Slot> = heap.get(arr_ref)?.fields.clone();
+    let n = i32::try_from(elems.len()).unwrap_or(0);
+    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
+    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(n);
+    heap.get_mut(stream_ref)?.fields.extend(elems);
+    Ok(Some(Slot::Reference(Some(stream_ref))))
+}
+/// Native: `Comparator.comparing(Function)Comparator` — creates a comparator by key extractor.
+/// Returns a `duke/util/ComparingComparator` with `fields[0]`=fn\_ref.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_comparator_comparing(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let fn_slot = extract_slot_arg(args, 0);
+    let r = heap.allocate("duke/util/ComparingComparator".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = fn_slot;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Arrays.sort(Object[])V` — natural order sort using `compareTo`.
+#[allow(clippy::too_many_lines)]
+pub(crate) fn native_arrays_sort_objects(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let len = heap.get(arr_ref)?.fields.len();
+    // Insertion sort with compareTo callbacks
+    for i in 1..len {
+        let mut j = i;
+        while j > 0 {
+            let a = heap.get(arr_ref)?.fields[j - 1];
+            let b = heap.get(arr_ref)?.fields[j];
+            let cmp = compare_slots_natural(a, b, heap, out, ops)?;
+            if cmp <= 0 {
+                break;
+            }
+            heap.write_field(arr_ref, j - 1, b)?;
+            heap.write_field(arr_ref, j, a)?;
+            j -= 1;
+        }
+    }
+    Ok(None)
+}
+/// Native: `Collections.singletonList(Object)List` — returns a one-element `ArrayList`.
+pub(crate) fn native_collections_singleton_list(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let element = extract_slot_arg(args, 0);
+    let r = heap.allocate("java/util/ArrayList".to_string(), 1);
+    native_arraylist_init(&[Slot::Reference(Some(r))], heap, out, control)?;
+    native_arraylist_add(&[Slot::Reference(Some(r)), element], heap, out, control)?;
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Collections.reverse(List)V` — reverses an `ArrayList` in-place.
+pub(crate) fn native_collections_reverse(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => return Ok(None),
+    };
+    // Elements are at fields[1..=size]; reverse that slice.
+    let fields = &mut heap.get_mut(list_ref)?.fields;
+    fields[1..=size].reverse();
+    Ok(None)
+}
+/// Native: `Collections.frequency(Collection, Object)I` — count occurrences of element.
+pub(crate) fn native_collections_frequency(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let coll_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let size = match heap.get(coll_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => return Ok(Some(Slot::Int(0))),
+    };
+    let fields = heap.get(coll_ref)?.fields[1..=size].to_vec();
+    let count = fields
+        .iter()
+        .filter(|s| slots_equal(s, &target, heap))
+        .count();
+    Ok(Some(Slot::Int(i32::try_from(count).unwrap_or(i32::MAX))))
+}
+/// Native: `Map$Entry.getKey()Object` — returns the key field.
+pub(crate) fn native_map_entry_get_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_first_field_arg(heap, this_ref)?;
+    Ok(Some(key))
+}
+/// Native: `Map$Entry.getValue()Object` — returns the value field.
+pub(crate) fn native_map_entry_get_value(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let val = extract_field_arg(heap, this_ref, 1)?;
+    Ok(Some(val))
+}
+/// Native: `HashMap.keySet()` — returns a new `HashSet` containing all keys.
+pub(crate) fn native_hashmap_key_set(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fields_len = heap.get(this_ref)?.fields.len();
+    let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(set_ref))], heap, out, control)?;
+    let mut i = 1usize;
+    while i < fields_len {
+        let key = heap.get(this_ref)?.fields[i];
+        native_hashset_add(&[Slot::Reference(Some(set_ref)), key], heap, out, control)?;
+        i += 2;
+    }
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+/// Native: `HashMap.values()` — returns a new `ArrayList` containing all values.
+/// `HashMap` fields: `[size, key0, val0, key1, val1, ...]`; values are at even indices 2, 4, 6, ...
+pub(crate) fn native_hashmap_values(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fields_len = heap.get(this_ref)?.fields.len();
+    let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
+    native_arraylist_init(&[Slot::Reference(Some(list_ref))], heap, out, control)?;
+    // Pairs start at index 1; values are at indices 2, 4, 6, ...
+    let mut i = 2usize;
+    while i < fields_len {
+        let val = heap.get(this_ref)?.fields[i];
+        native_arraylist_add(&[Slot::Reference(Some(list_ref)), val], heap, out, control)?;
+        i += 2;
+    }
+    Ok(Some(Slot::Reference(Some(list_ref))))
+}
+/// Native: `HashMap.entrySet()` — returns a new `HashSet` of `java/util/Map$Entry` objects.
+pub(crate) fn native_hashmap_entry_set(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fields_len = heap.get(this_ref)?.fields.len();
+    let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(set_ref))], heap, out, control)?;
+    let mut i = 1usize;
+    while i + 1 < fields_len {
+        let key = heap.get(this_ref)?.fields[i];
+        let val = heap.get(this_ref)?.fields[i + 1];
+        let entry_ref = heap.allocate("java/util/Map$Entry".to_string(), 2);
+        {
+            let entry_obj = heap.get_mut(entry_ref)?;
+            entry_obj.fields[0] = key;
+            entry_obj.fields[1] = val;
+        }
+        native_hashset_add(
+            &[
+                Slot::Reference(Some(set_ref)),
+                Slot::Reference(Some(entry_ref)),
+            ],
+            heap,
+            out,
+            control,
+        )?;
+        i += 2;
+    }
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+/// Native: `ArrayList.<init>()V` — initializes with size=0.
+pub(crate) fn native_arraylist_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `ArrayList.<init>(I)V` — initializes with size=0, ignoring the initial capacity.
+pub(crate) fn native_arraylist_init_with_capacity(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `ArrayList.<init>(Collection)V` — copies elements from another `ArrayList`/collection.
+pub(crate) fn native_arraylist_init_from_collection(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let Some(Slot::Reference(Some(src_ref))) = args.get(1).copied() else {
+        heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+        return Ok(None);
+    };
+    // Copy size and elements from source collection (ArrayList layout: fields[0]=size, fields[1..]=elems)
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let src_size = match src_fields.first() {
+        Some(Slot::Int(n)) => *n,
+        _ => 0,
+    };
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(src_size);
+    for elem in src_fields
+        .into_iter()
+        .skip(1)
+        .take(usize::try_from(src_size).unwrap_or(0))
+    {
+        heap.get_mut(this_ref)?.fields.push(elem);
+    }
+    Ok(None)
+}
+/// Native: `ArrayList.add(Object)Z` — appends element, returns true.
+pub(crate) fn native_arraylist_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let element = extract_slot_arg(args, 1);
+    let obj = heap.get_mut(this_ref)?;
+    match obj.fields.first_mut() {
+        Some(Slot::Int(sz)) => *sz += 1,
+        _ => return Err(Error::NullPointerException), // shouldn't happen; init sets fields[0]=Int(0)
+    }
+    obj.fields.push(element);
+    Ok(Some(Slot::Int(1))) // boolean true
+}
+/// Native: `ArrayList.get(I)Object` — returns element at index.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_arraylist_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let idx = extract_int_arg(args, 1)? as usize;
+    let obj = heap.get(this_ref)?;
+    obj.fields.get(idx + 1).map_or_else(
+        || {
+            Err(Error::JavaException {
+                class_name: "java/lang/ArrayIndexOutOfBoundsException".to_string(),
+            })
+        },
+        |slot| Ok(Some(*slot)),
+    )
+}
+/// Native: `ArrayList.size()I`
+pub(crate) fn native_arraylist_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let obj = heap.get(this_ref)?;
+    match obj.fields.first() {
+        Some(Slot::Int(sz)) => Ok(Some(Slot::Int(*sz))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+/// Native: `ArrayList.iterator()Iterator` — creates an `ArrayListIterator`.
+pub(crate) fn native_arraylist_iterator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    // fields[0]=list_ref, fields[1]=cursor, fields[2]=last_returned (-1 = none)
+    let iter_ref = heap.allocate("duke/util/ArrayListIterator".to_string(), 3);
+    {
+        let iter_obj = heap.get_mut(iter_ref)?;
+        iter_obj.fields[0] = Slot::Reference(Some(this_ref));
+        iter_obj.fields[1] = Slot::Int(0);
+        iter_obj.fields[2] = Slot::Int(-1);
+    }
+    Ok(Some(Slot::Reference(Some(iter_ref))))
+}
+/// Native: `Collection.toArray()` — copies Duke-backed collection elements into `Object[]`.
+pub(crate) fn native_collection_to_array(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elements = collection_elements_from_ref(heap, this_ref)?;
+    let array_ref = allocate_reference_array_from_slots(heap, "[Ljava/lang/Object;", &elements)?;
+    Ok(Some(Slot::Reference(Some(array_ref))))
+}
+/// Native: `Collection.toArray(Object[])` — preserves the requested array type.
+pub(crate) fn native_collection_to_array_with_seed_array(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let seed_array_ref = extract_ref_arg(args, 1)?;
+    let elements = collection_elements_from_ref(heap, this_ref)?;
+    let seed_class_name = heap.get(seed_array_ref)?.class_name.clone();
+    let seed_len = heap.get(seed_array_ref)?.fields.len();
+
+    if seed_len < elements.len() {
+        let new_array_ref = allocate_reference_array_from_slots(heap, &seed_class_name, &elements)?;
+        return Ok(Some(Slot::Reference(Some(new_array_ref))));
+    }
+
+    let seed_array = heap.get_mut(seed_array_ref)?;
+    for slot in &mut seed_array.fields {
+        *slot = Slot::Reference(None);
+    }
+    for (idx, element) in elements.iter().enumerate() {
+        seed_array.fields[idx] = *element;
+    }
+    Ok(Some(Slot::Reference(Some(seed_array_ref))))
+}
+/// Native: `Collections.sort(List)V` — delegates to the list's sort(null) method.
+///
+/// `Collections.sort(list)` is compiled by javac as
+/// `invokestatic java/util/Collections.sort:(Ljava/util/List;)V`.
+/// We forward to the runtime class's `sort(Comparator=null)`, which for an
+/// `ArrayList` performs the insertion-sort-with-compareTo callback.
+pub(crate) fn native_collections_sort(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    // args[0] = List ref
+    let list_ref = extract_ref_arg(args, 0)?;
+    // Dispatch on the actual runtime class so any List implementation works.
+    let class_name = heap.get(list_ref)?.class_name.clone();
+    ops.invoke(
+        heap,
+        output,
+        &class_name,
+        "sort",
+        SORT_COMPARATOR_DESC,
+        vec![Slot::Reference(Some(list_ref)), Slot::Reference(None)],
+    )?;
+    Ok(None)
+}
+pub(crate) fn native_service_loader_load(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    allocate_service_loader(args, heap, ops, None)
+}
+pub(crate) fn native_service_loader_load_with_loader(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    allocate_service_loader(args, heap, ops, Some(1))
+}
+pub(crate) fn native_service_loader_iterator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (service_class_slot, loader_slot, provider_slots) = {
+        let service_loader = heap.get(this_ref)?;
+        let count = match service_loader.fields.get(SERVICE_LOADER_COUNT_FIELD) {
+            Some(Slot::Int(value)) if *value > 0 => usize::try_from(*value).unwrap_or(0),
+            _ => 0,
+        };
+        let service_class_slot = service_loader.fields[SERVICE_LOADER_SERVICE_CLASS_FIELD];
+        let loader_slot = service_loader.fields[SERVICE_LOADER_LOADER_FIELD];
+        let provider_slots: Vec<Slot> = service_loader
+            .fields
+            .iter()
+            .skip(SERVICE_LOADER_PROVIDERS_START)
+            .take(count)
+            .copied()
+            .collect();
+        (service_class_slot, loader_slot, provider_slots)
+    };
+
+    let iter_ref = heap.allocate(
+        "duke/util/ServiceLoaderIterator".to_string(),
+        SERVICE_ITER_PROVIDERS_START + provider_slots.len(),
+    );
+    {
+        let iter = heap.get_mut(iter_ref)?;
+        iter.fields[SERVICE_ITER_SERVICE_CLASS_FIELD] = service_class_slot;
+        iter.fields[SERVICE_ITER_LOADER_FIELD] = loader_slot;
+        iter.fields[SERVICE_ITER_INDEX_FIELD] = Slot::Int(0);
+        iter.fields[SERVICE_ITER_COUNT_FIELD] =
+            Slot::Int(i32::try_from(provider_slots.len()).unwrap_or(i32::MAX));
+    }
+    for (idx, slot) in provider_slots.into_iter().enumerate() {
+        heap.write_field(iter_ref, SERVICE_ITER_PROVIDERS_START + idx, slot)?;
+    }
+    Ok(Some(Slot::Reference(Some(iter_ref))))
+}
+pub(crate) fn native_service_loader_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let Some(Slot::Reference(Some(iter_ref))) =
+        native_service_loader_iterator(args, heap, output, control)?
+    else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    let mut providers = Vec::new();
+    loop {
+        let (index, count) = service_iterator_index_and_count(heap, iter_ref)?;
+        if index >= count {
+            break;
+        }
+        let provider_ref = instantiate_service_provider(heap, output, ops, iter_ref)?;
+        providers.push(provider_ref);
+    }
+
+    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
+    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(i32::try_from(providers.len()).unwrap_or(0));
+    for provider_ref in providers {
+        heap.get_mut(stream_ref)?
+            .fields
+            .push(Slot::Reference(Some(provider_ref)));
+    }
+    Ok(Some(Slot::Reference(Some(stream_ref))))
+}
+/// Native: `ArrayList.remove(I)Object` — removes element at index, returns it.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_arraylist_remove_at(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let idx = extract_int_arg(args, 1)?;
+    if idx < 0 {
+        return Err(Error::JavaException {
+            class_name: "java/lang/IndexOutOfBoundsException".to_string(),
+        });
+    }
+    let idx = idx as usize;
+    let len = heap.get(this_ref)?.fields.len();
+    // fields[0]=size, elements start at 1; idx is 0-based element index → field index = idx+1
+    if idx + 1 >= len {
+        return Err(Error::JavaException {
+            class_name: "java/lang/IndexOutOfBoundsException".to_string(),
+        });
+    }
+    let removed = heap.get(this_ref)?.fields[idx + 1];
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields.remove(idx + 1);
+    if let Some(Slot::Int(sz)) = obj.fields.first_mut() {
+        *sz -= 1;
+    }
+    Ok(Some(removed))
+}
+/// Native: `ArrayList.remove(Object)Z` — removes first occurrence, returns true if found.
+pub(crate) fn native_arraylist_remove_obj(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let mut found = None;
+    for (i, slot) in heap.get(this_ref)?.fields.iter().enumerate().skip(1) {
+        if slots_equal(slot, &target, heap) {
+            found = Some(i);
+            break;
+        }
+    }
+    if let Some(field_idx) = found {
+        let obj = heap.get_mut(this_ref)?;
+        obj.fields.remove(field_idx);
+        if let Some(Slot::Int(sz)) = obj.fields.first_mut() {
+            *sz -= 1;
+        }
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+/// Native: `ArrayList.contains(Object)Z` — returns 1 if element is present.
+pub(crate) fn native_arraylist_contains(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let mut found = false;
+    for slot in heap.get(this_ref)?.fields.iter().skip(1) {
+        if slots_equal(slot, &target, heap) {
+            found = true;
+            break;
+        }
+    }
+    Ok(Some(Slot::Int(i32::from(found))))
+}
+/// Native: `ArrayList.clear()V` — removes all elements.
+pub(crate) fn native_arraylist_clear(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields.truncate(1);
+    obj.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `ArrayList.isEmpty()Z` — returns 1 if size is 0.
+pub(crate) fn native_arraylist_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let is_empty = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(sz)) => *sz == 0,
+        _ => true,
+    };
+    Ok(Some(Slot::Int(i32::from(is_empty))))
+}
+/// Native: `ArrayList.set(I,Object)Object` — replaces element at index, returns old value.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_arraylist_set(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let idx = extract_int_arg(args, 1)?;
+    let value = extract_slot_arg(args, 2);
+    if idx < 0 {
+        return Err(Error::JavaException {
+            class_name: "java/lang/IndexOutOfBoundsException".to_string(),
+        });
+    }
+    let field_idx = idx as usize + 1;
+    let old = *heap
+        .get(this_ref)?
+        .fields
+        .get(field_idx)
+        .ok_or_else(|| Error::JavaException {
+            class_name: "java/lang/IndexOutOfBoundsException".to_string(),
+        })?;
+    heap.get_mut(this_ref)?.fields[field_idx] = value;
+    Ok(Some(old))
+}
+/// Native: `ArrayList.indexOf(Object)I` — returns first index of element, or -1.
+pub(crate) fn native_arraylist_index_of(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let len = heap.get(this_ref)?.fields.len();
+    for i in 1..len {
+        let slot = heap.get(this_ref)?.fields[i];
+        if slots_equal(&slot, &target, heap) {
+            return Ok(Some(Slot::Int(i32::try_from(i - 1).unwrap_or(i32::MAX))));
+        }
+    }
+    Ok(Some(Slot::Int(-1)))
+}
+/// Native: `ArrayList.lastIndexOf(Object)I` — last occurrence, or -1.
+pub(crate) fn native_arraylist_last_index_of(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let len = heap.get(this_ref)?.fields.len();
+    if len > 1 {
+        for i in (1..len).rev() {
+            let slot = heap.get(this_ref)?.fields[i];
+            if slots_equal(&slot, &target, heap) {
+                return Ok(Some(Slot::Int(i32::try_from(i - 1).unwrap_or(i32::MAX))));
+            }
+        }
+    }
+    Ok(Some(Slot::Int(-1)))
+}
+/// Native: `ArrayList.add(I,Object)V` — inserts element at index, shifting others right.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_arraylist_add_at(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let idx = extract_int_arg(args, 1)?;
+    let element = extract_slot_arg(args, 2);
+    if idx < 0 {
+        return Err(Error::JavaException {
+            class_name: "java/lang/IndexOutOfBoundsException".to_string(),
+        });
+    }
+    let field_idx = idx as usize + 1;
+    let obj = heap.get_mut(this_ref)?;
+    let len = obj.fields.len();
+    if field_idx > len {
+        return Err(Error::JavaException {
+            class_name: "java/lang/IndexOutOfBoundsException".to_string(),
+        });
+    }
+    obj.fields.insert(field_idx, element);
+    if let Some(Slot::Int(sz)) = obj.fields.first_mut() {
+        *sz += 1;
+    }
+    Ok(None)
+}
+/// Native: `HashMap.putIfAbsent(K,V)Object` — inserts only if key is absent; returns existing or null.
+pub(crate) fn native_hashmap_put_if_absent(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let val = extract_slot_arg(args, 2);
+    let i_opt = find_hashmap_entry_index(&heap.get(this_ref)?.fields, &key, heap);
+    if let Some(i) = i_opt {
+        // Key already present — return existing value.
+        return Ok(Some(heap.get(this_ref)?.fields[i + 1]));
+    }
+    // Key absent — insert and return null.
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(Slot::Int(sz)) = obj.fields.first_mut() {
+        *sz += 1;
+    }
+    obj.fields.push(key);
+    obj.fields.push(val);
+    Ok(Some(Slot::Reference(None)))
+}
+/// Native: `HashMap.clear()V` — removes all entries.
+pub(crate) fn native_hashmap_clear(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields.truncate(1);
+    obj.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `HashMap.containsValue(Object)Z` — returns 1 if any entry has this value.
+pub(crate) fn native_hashmap_contains_value(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let mut i = 2usize;
+    loop {
+        // Evaluate len inside the loop since heap access can't be held across slots_equal
+        let len = heap.get(this_ref)?.fields.len();
+        if i >= len {
+            break;
+        }
+        let slot = heap.get(this_ref)?.fields[i];
+        if slots_equal(&slot, &target, heap) {
+            return Ok(Some(Slot::Int(1)));
+        }
+        i += 2;
+    }
+    Ok(Some(Slot::Int(0)))
+}
+/// Native: `Arrays.fill(int[], int)` — fills all elements with val.
+pub(crate) fn native_arrays_fill_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => Slot::Int(*v),
+        _ => Slot::Int(0),
+    };
+    let obj = heap.get_mut(arr_ref)?;
+    for slot in &mut obj.fields {
+        *slot = val;
+    }
+    Ok(None)
+}
+/// Native: `Arrays.fill(Object[], Object)` — fills all elements with val.
+pub(crate) fn native_arrays_fill_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let val = extract_slot_arg(args, 1);
+    let obj = heap.get_mut(arr_ref)?;
+    for slot in &mut obj.fields {
+        *slot = val;
+    }
+    Ok(None)
+}
+/// Native: `Arrays.copyOf(int[], int)` — copies to new int[] of given length.
+pub(crate) fn native_arrays_copyof_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let src_ref = extract_ref_arg(args, 0)?;
+    let new_len = match args.get(1) {
+        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
+        Some(Slot::Int(n)) => return Err(Error::NegativeArraySize { size: *n }),
+        _ => 0,
+    };
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let dst_ref = heap.allocate("[I".to_string(), new_len);
+    let dst = heap.get_mut(dst_ref)?;
+    for i in 0..new_len {
+        dst.fields[i] = src_fields.get(i).copied().unwrap_or(Slot::Int(0));
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+/// Native: `Arrays.copyOf(Object[], int)` — copies to new Object[] of given length.
+pub(crate) fn native_arrays_copyof_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let src_ref = extract_ref_arg(args, 0)?;
+    let new_len = match args.get(1) {
+        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
+        Some(Slot::Int(n)) => return Err(Error::NegativeArraySize { size: *n }),
+        _ => 0,
+    };
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let dst_ref = heap.allocate("[Ljava/lang/Object;".to_string(), new_len);
+    let dst = heap.get_mut(dst_ref)?;
+    for i in 0..new_len {
+        dst.fields[i] = src_fields.get(i).copied().unwrap_or(Slot::Reference(None));
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+/// Native: `Arrays.sort(int[])` — sorts fields in place.
+pub(crate) fn native_arrays_sort_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let obj = heap.get_mut(arr_ref)?;
+    obj.fields.sort_by(|a, b| match (a, b) {
+        (Slot::Int(x), Slot::Int(y)) => x.cmp(y),
+        _ => std::cmp::Ordering::Equal,
+    });
+    Ok(None)
+}
+/// Native: `Arrays.equals(int[], int[])boolean` — element-wise equality.
+pub(crate) fn native_arrays_equals_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a_ref = extract_ref_arg(args, 0)?;
+    let b_ref = extract_ref_arg(args, 1)?;
+    let a_len = heap.get(a_ref)?.fields.len();
+    let b_len = heap.get(b_ref)?.fields.len();
+    if a_len != b_len {
+        return Ok(Some(Slot::Int(0)));
+    }
+    for i in 0..a_len {
+        let x = heap.get(a_ref)?.fields[i];
+        let y = heap.get(b_ref)?.fields[i];
+        match (x, y) {
+            (Slot::Int(a), Slot::Int(b)) if a == b => {}
+            _ => return Ok(Some(Slot::Int(0))),
+        }
+    }
+    Ok(Some(Slot::Int(1)))
+}
+/// Native: `Arrays.copyOfRange(int[], int, int)int[]` — slice of int array, zero-padded.
+pub(crate) fn native_arrays_copy_of_range_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let src_ref = extract_ref_arg(args, 0)?;
+    let from = match args.get(1) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let to = match args.get(2) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let new_len = to.saturating_sub(from);
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let dst_ref = heap.allocate("[I".to_string(), new_len);
+    for i in 0..new_len {
+        heap.get_mut(dst_ref)?.fields[i] =
+            src_fields.get(from + i).copied().unwrap_or(Slot::Int(0));
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+/// Native: `Arrays.copyOfRange(Object[], int, int)Object[]` — slice of reference array.
+pub(crate) fn native_arrays_copy_of_range_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let src_ref = extract_ref_arg(args, 0)?;
+    let from = match args.get(1) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let to = match args.get(2) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let new_len = to.saturating_sub(from);
+    let (src_class, src_fields) = {
+        let obj = heap.get(src_ref)?;
+        (obj.class_name.clone(), obj.fields.clone())
+    };
+    let dst_ref = heap.allocate(src_class, new_len);
+    for i in 0..new_len {
+        heap.get_mut(dst_ref)?.fields[i] = src_fields
+            .get(from + i)
+            .copied()
+            .unwrap_or(Slot::Reference(None));
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+/// Native: `ArrayList.subList(int, int)List` — returns a new `ArrayList` with the sub-range.
+pub(crate) fn native_arraylist_sub_list(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let from = match args.get(1) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let to = match args.get(2) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let new_len = to.saturating_sub(from);
+    // ArrayList layout: fields[0]=size, fields[1..]=elements
+    let src_elems: Vec<Slot> = {
+        let fields = &heap.get(this_ref)?.fields;
+        fields
+            .iter()
+            .skip(1 + from)
+            .take(new_len)
+            .copied()
+            .collect()
+    };
+    let sub_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
+    heap.get_mut(sub_ref)?.fields[0] = Slot::Int(i32::try_from(new_len).unwrap_or(0));
+    heap.get_mut(sub_ref)?.fields.extend(src_elems);
+    Ok(Some(Slot::Reference(Some(sub_ref))))
+}
+/// Native: `ArrayList.removeIf(Predicate)Z` — removes all elements where predicate returns true.
+pub(crate) fn native_arraylist_remove_if(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fn_slot = extract_slot_arg(args, 1);
+    let Slot::Reference(Some(fn_ref)) = fn_slot else {
+        return Ok(Some(Slot::Int(0)));
+    };
+    let fn_class = heap.get(fn_ref)?.class_name.clone();
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(this_ref)?.fields[1..=size].to_vec();
+    let mut kept = Vec::with_capacity(elems.len());
+    let mut removed = false;
+    for elem in elems {
+        let result = ops
+            .invoke(
+                heap,
+                out,
+                &fn_class,
+                "test",
+                "(Ljava/lang/Object;)Z",
+                vec![fn_slot, elem],
+            )?
+            .unwrap_or(Slot::Int(0));
+        match result {
+            Slot::Int(1) => {
+                removed = true;
+            } // predicate true → remove
+            _ => kept.push(elem),
+        }
+    }
+    let new_size = i32::try_from(kept.len()).unwrap_or(0);
+    heap.get_mut(this_ref)?.fields.truncate(1);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(new_size);
+    heap.get_mut(this_ref)?.fields.extend(kept);
+    Ok(Some(Slot::Int(i32::from(removed))))
+}
+/// Native: `Collections.binarySearch(List, T)I` — binary search on sorted `ArrayList`.
+pub(crate) fn native_collections_binary_search(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(list_ref)?.fields[1..=size].to_vec();
+    let mut lo: i64 = 0;
+    let mut hi: i64 = i64::try_from(elems.len()).unwrap_or(0) - 1;
+    while lo <= hi {
+        let mid = lo + (hi - lo) / 2; // avoid overflow via i64 midpoint
+        let mid_elem = elems[usize::try_from(mid).unwrap_or(0)];
+        let cmp = compare_slots_natural(mid_elem, key, heap, out, ops)?;
+        match cmp.cmp(&0) {
+            std::cmp::Ordering::Equal => {
+                return Ok(Some(Slot::Int(i32::try_from(mid).unwrap_or(0))));
+            }
+            std::cmp::Ordering::Less => lo = mid + 1,
+            std::cmp::Ordering::Greater => hi = mid - 1,
+        }
+    }
+    // Return -(insertion point) - 1
+    let insertion = i32::try_from(lo).unwrap_or(0);
+    Ok(Some(Slot::Int(-(insertion + 1))))
+}
+/// Native: `Arrays.asList(Object[])List` — wraps a reference array as an `ArrayList`.
+pub(crate) fn native_arrays_as_list(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let arr_len = heap.get(arr_ref)?.fields.len();
+    // Create a new ArrayList (1 field slot for size counter) and populate it.
+    let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
+    native_arraylist_init(&[Slot::Reference(Some(list_ref))], heap, out, control)?;
+    for i in 0..arr_len {
+        let elem = extract_field_arg(heap, arr_ref, i)?;
+        native_arraylist_add(&[Slot::Reference(Some(list_ref)), elem], heap, out, control)?;
+    }
+    Ok(Some(Slot::Reference(Some(list_ref))))
+}
+/// Native: `Random.<init>()V` — seed from current time.
+#[allow(clippy::unnecessary_wraps, clippy::cast_possible_wrap)]
+pub(crate) fn native_random_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let nanos = u64::from(
+        std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .map_or(0, |d| d.subsec_nanos()),
+    );
+    let initial = (nanos ^ RANDOM_MULTIPLIER) & RANDOM_MASK;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Long(initial as i64); // safe: mask ensures < 2^48
+    Ok(None)
+}
+/// Native: `Random.<init>(J)V` — seed with explicit long value.
+#[allow(
+    clippy::unnecessary_wraps,
+    clippy::cast_sign_loss,
+    clippy::cast_possible_wrap
+)]
+pub(crate) fn native_random_init_seed(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let seed = match args.get(1).copied() {
+        Some(Slot::Long(v)) => v as u64,
+        Some(Slot::Int(v)) => v as u64,
+        _ => 0,
+    };
+    let initial = (seed ^ RANDOM_MULTIPLIER) & RANDOM_MASK;
+    heap.get_mut(this_ref)?.fields[0] = Slot::Long(initial as i64); // safe: < 2^48
+    Ok(None)
+}
+/// Native: `Random.nextInt()I` — full-range random int.
+pub(crate) fn native_random_next_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, v) = random_step(heap, this_ref, 32)?;
+    Ok(Some(Slot::Int(v)))
+}
+/// Native: `Random.nextInt(I)I` — bounded random int [0, bound).
+#[allow(
+    clippy::cast_sign_loss,
+    clippy::cast_possible_wrap,
+    clippy::cast_possible_truncation
+)]
+pub(crate) fn native_random_next_int_bound(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let bound = extract_int_arg(args, 1)?;
+    if bound <= 0 {
+        return Err(duke_runtime::Error::JavaException {
+            class_name: "java/lang/IllegalArgumentException".to_string(),
+        });
+    }
+    let bound_u = bound as u32;
+    // Java rejection-sampling to avoid modulo bias
+    loop {
+        let (_, bits) = random_step(heap, this_ref, 31)?;
+        let bits_u = bits as u32; // bits from next(31) are always non-negative
+        let val = bits_u % bound_u;
+        if bits_u.wrapping_sub(val).wrapping_add(bound_u - 1) < u32::MAX {
+            return Ok(Some(Slot::Int(val as i32))); // val < bound <= i32::MAX
+        }
+    }
+}
+/// Native: `Random.nextLong()J` — 64-bit random long (two 32-bit calls).
+pub(crate) fn native_random_next_long(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, hi) = random_step(heap, this_ref, 32)?;
+    let (_, lo) = random_step(heap, this_ref, 32)?;
+    let v = (i64::from(hi) << 32) + i64::from(lo);
+    Ok(Some(Slot::Long(v)))
+}
+/// Native: `Random.nextDouble()D` — uniform [0.0, 1.0).
+#[allow(clippy::cast_precision_loss)]
+pub(crate) fn native_random_next_double(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, hi) = random_step(heap, this_ref, 26)?;
+    let (_, lo) = random_step(heap, this_ref, 27)?;
+    // Java spec: ((long)(next(26)) << 27) + next(27)) / (double)(1L << 53)
+    let combined = (i64::from(hi) << 27) + i64::from(lo);
+    let v = combined as f64 / (1u64 << 53) as f64;
+    Ok(Some(Slot::Double(v)))
+}
+/// Native: `Random.nextFloat()F` — uniform [0.0, 1.0) as float.
+#[allow(clippy::cast_precision_loss)]
+pub(crate) fn native_random_next_float(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, bits) = random_step(heap, this_ref, 24)?;
+    // next(24) is non-negative, safe to cast to f32
+    let v = bits as f32 / (1u32 << 24) as f32;
+    Ok(Some(Slot::Float(v)))
+}
+/// Native: `Random.nextBoolean()Z` — random boolean.
+pub(crate) fn native_random_next_boolean(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, v) = random_step(heap, this_ref, 1)?;
+    Ok(Some(Slot::Int(v)))
+}
+/// Native: `UUID.<init>(long,long)V` — stores the two canonical 64-bit halves.
+pub(crate) fn native_uuid_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let msb = extract_long_arg(args, 1)?;
+    let lsb = extract_long_arg(args, 2)?;
+    write_uuid_bits(heap, this_ref, msb, lsb)?;
+    Ok(None)
+}
+/// Native: `UUID.randomUUID()UUID` — RFC 4122 version-4 UUID from host CSPRNG.
+pub(crate) fn native_uuid_random_uuid(
+    _args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let mut bytes = [0_u8; 16];
+    getrandom::fill(&mut bytes).map_err(|_| Error::JavaException {
+        class_name: "java/lang/InternalError".to_string(),
+    })?;
+    bytes[6] = (bytes[6] & 0x0f) | 0x40;
+    bytes[8] = (bytes[8] & 0x3f) | 0x80;
+    let (msb, lsb) = uuid_bits_from_bytes(&bytes);
+    let uuid_ref = allocate_uuid(heap, msb, lsb)?;
+    Ok(Some(Slot::Reference(Some(uuid_ref))))
+}
+/// Native: `UUID.nameUUIDFromBytes(byte[])UUID` — RFC 4122 version-3 MD5 UUID.
+pub(crate) fn native_uuid_name_uuid_from_bytes(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let input_ref = extract_ref_arg(args, 0)?;
+    let input = byte_array_from_ref(heap, input_ref)?;
+    let digest = compute_message_digest("MD5", &input)?;
+    let mut bytes = [0_u8; 16];
+    bytes.copy_from_slice(&digest[..16]);
+    bytes[6] = (bytes[6] & 0x0f) | 0x30;
+    bytes[8] = (bytes[8] & 0x3f) | 0x80;
+    let (msb, lsb) = uuid_bits_from_bytes(&bytes);
+    let uuid_ref = allocate_uuid(heap, msb, lsb)?;
+    Ok(Some(Slot::Reference(Some(uuid_ref))))
+}
+/// Native: `UUID.fromString(String)UUID` — parses canonical 8-4-4-4-12 UUID text.
+pub(crate) fn native_uuid_from_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let text_ref = extract_ref_arg(args, 0)?;
+    let text = string_value_from_ref(heap, text_ref)?;
+    let (msb, lsb) = parse_uuid_string(&text)?;
+    let uuid_ref = allocate_uuid(heap, msb, lsb)?;
+    Ok(Some(Slot::Reference(Some(uuid_ref))))
+}
+/// Native: `UUID.getMostSignificantBits()long`.
+pub(crate) fn native_uuid_get_most_significant_bits(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (msb, _) = uuid_bits_from_ref(heap, this_ref)?;
+    Ok(Some(Slot::Long(msb)))
+}
+/// Native: `UUID.getLeastSignificantBits()long`.
+pub(crate) fn native_uuid_get_least_significant_bits(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, lsb) = uuid_bits_from_ref(heap, this_ref)?;
+    Ok(Some(Slot::Long(lsb)))
+}
+/// Native: `UUID.version()int`.
+pub(crate) fn native_uuid_version(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (msb, _) = uuid_bits_from_ref(heap, this_ref)?;
+    let version = i32::try_from((msb.cast_unsigned() >> 12) & 0x0f)
+        .expect("UUID version nibble fits in i32");
+    Ok(Some(Slot::Int(version)))
+}
+/// Native: `UUID.variant()int`.
+pub(crate) fn native_uuid_variant(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (_, lsb) = uuid_bits_from_ref(heap, this_ref)?;
+    let lsb = lsb.cast_unsigned();
+    let variant = if (lsb >> 63) == 0 {
+        0
+    } else if (lsb >> 62) == 0b10 {
+        2
+    } else if (lsb >> 61) == 0b110 {
+        6
+    } else {
+        7
+    };
+    Ok(Some(Slot::Int(variant)))
+}
+/// Native: `UUID.toString()String`.
+pub(crate) fn native_uuid_to_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (msb, lsb) = uuid_bits_from_ref(heap, this_ref)?;
+    let string_ref = heap.allocate_string(uuid_to_string(msb, lsb));
+    Ok(Some(Slot::Reference(Some(string_ref))))
+}
+/// Native: `UUID.equals(Object)boolean`.
+pub(crate) fn native_uuid_equals(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let Ok(other_ref) = extract_ref_arg(args, 1) else {
+        return Ok(Some(Slot::Int(0)));
+    };
+    let this_bits = uuid_bits_from_ref(heap, this_ref)?;
+    let other = heap.get(other_ref)?;
+    let equal = other.class_name == "java/util/UUID"
+        && uuid_bits_from_object(other).is_some_and(|other_bits| other_bits == this_bits);
+    Ok(Some(Slot::Int(i32::from(equal))))
+}
+/// Native: `UUID.hashCode()int`.
+pub(crate) fn native_uuid_hash_code(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let (msb, lsb) = uuid_bits_from_ref(heap, this_ref)?;
+    let hilo = msb.cast_unsigned() ^ lsb.cast_unsigned();
+    let high = u32::try_from(hilo >> 32).expect("upper 32 bits fit in u32");
+    let low = u32::try_from(hilo & 0xffff_ffff).expect("lower 32 bits fit in u32");
+    Ok(Some(Slot::Int((high ^ low).cast_signed())))
+}
+/// Native: `UUID.compareTo(UUID)int`.
+pub(crate) fn native_uuid_compare_to(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let other_ref = extract_ref_arg(args, 1)?;
+    let this_bits = uuid_bits_from_ref(heap, this_ref)?;
+    let other_bits = uuid_bits_from_ref(heap, other_ref)?;
+    let ordering = this_bits
+        .0
+        .cmp(&other_bits.0)
+        .then_with(|| this_bits.1.cmp(&other_bits.1));
+    Ok(Some(Slot::Int(match ordering {
+        std::cmp::Ordering::Less => -1,
+        std::cmp::Ordering::Equal => 0,
+        std::cmp::Ordering::Greater => 1,
+    })))
+}
+/// Native: version-1 UUID accessors are intentionally deferred for v3/v4 UUIDs.
+pub(crate) fn native_uuid_unsupported_version1_accessor(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    Err(Error::JavaException {
+        class_name: "java/lang/UnsupportedOperationException".to_string(),
+    })
+}
+/// Native: `Optional.filter(Predicate)Optional` — keeps value only if predicate passes.
+pub(crate) fn native_optional_filter(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let opt_ref = extract_ref_arg(args, 0)?;
+    let pred_slot = extract_slot_arg(args, 1);
+    let value = extract_first_field_arg(heap, opt_ref)?;
+    let result_ref = heap.allocate("java/util/Optional".to_string(), 1);
+    if matches!(value, Slot::Reference(None)) {
+        heap.get_mut(result_ref)?.fields[0] = Slot::Reference(None);
+        return Ok(Some(Slot::Reference(Some(result_ref))));
+    }
+    let Slot::Reference(Some(pred_ref)) = pred_slot else {
+        heap.get_mut(result_ref)?.fields[0] = value;
+        return Ok(Some(Slot::Reference(Some(result_ref))));
+    };
+    let pred_class = heap.get(pred_ref)?.class_name.clone();
+    let test_result = ops.invoke(
+        heap,
+        out,
+        &pred_class,
+        "test",
+        "(Ljava/lang/Object;)Z",
+        vec![pred_slot, value],
+    )?;
+    let passes = matches!(test_result, Some(Slot::Int(n)) if n != 0);
+    let stored = if passes { value } else { Slot::Reference(None) };
+    heap.get_mut(result_ref)?.fields[0] = stored;
+    Ok(Some(Slot::Reference(Some(result_ref))))
+}
+/// Native: `Optional.flatMap(Function)Optional` — maps value to Optional if present, flattens.
+pub(crate) fn native_optional_flat_map(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let opt_ref = extract_ref_arg(args, 0)?;
+    let fn_slot = extract_slot_arg(args, 1);
+    let value = extract_first_field_arg(heap, opt_ref)?;
+    if matches!(value, Slot::Reference(None)) {
+        let empty_ref = heap.allocate("java/util/Optional".to_string(), 1);
+        heap.get_mut(empty_ref)?.fields[0] = Slot::Reference(None);
+        return Ok(Some(Slot::Reference(Some(empty_ref))));
+    }
+    let Slot::Reference(Some(fn_ref)) = fn_slot else {
+        let empty_ref = heap.allocate("java/util/Optional".to_string(), 1);
+        heap.get_mut(empty_ref)?.fields[0] = Slot::Reference(None);
+        return Ok(Some(Slot::Reference(Some(empty_ref))));
+    };
+    let fn_class = heap.get(fn_ref)?.class_name.clone();
+    // The function returns an Optional — return it directly (flat, not wrapped again)
+    let result = ops.invoke(
+        heap,
+        out,
+        &fn_class,
+        "apply",
+        "(Ljava/lang/Object;)Ljava/lang/Object;",
+        vec![fn_slot, value],
+    )?;
+    Ok(Some(result.unwrap_or(Slot::Reference(None))))
+}
+/// Native: `Optional.ifPresent(Consumer)V` — invokes consumer if value is present.
+pub(crate) fn native_optional_if_present(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let opt_ref = extract_ref_arg(args, 0)?;
+    let consumer_slot = extract_slot_arg(args, 1);
+    let value = extract_first_field_arg(heap, opt_ref)?;
+    if let (Slot::Reference(Some(_)), Slot::Reference(Some(consumer_ref))) = (value, consumer_slot)
+    {
+        let consumer_class = heap.get(consumer_ref)?.class_name.clone();
+        ops.invoke(
+            heap,
+            out,
+            &consumer_class,
+            "accept",
+            "(Ljava/lang/Object;)V",
+            vec![consumer_slot, value],
+        )?;
+    }
+    Ok(None)
+}
+/// Native: `Optional.orElseGet(Supplier)Object` — calls supplier if empty.
+pub(crate) fn native_optional_or_else_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let opt_ref = extract_ref_arg(args, 0)?;
+    let supplier_slot = extract_slot_arg(args, 1);
+    let value = extract_first_field_arg(heap, opt_ref)?;
+    if !matches!(value, Slot::Reference(None)) {
+        return Ok(Some(value));
+    }
+    let Slot::Reference(Some(supplier_ref)) = supplier_slot else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    let supplier_class = heap.get(supplier_ref)?.class_name.clone();
+    let result = ops.invoke(
+        heap,
+        out,
+        &supplier_class,
+        "get",
+        "()Ljava/lang/Object;",
+        vec![supplier_slot],
+    )?;
+    Ok(Some(result.unwrap_or(Slot::Reference(None))))
+}
+/// Native: `HashMap.compute(K, BiFunction)V` — compute new value from old (possibly null).
+pub(crate) fn native_hashmap_compute(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let fn_slot = extract_slot_arg(args, 2);
+    let Slot::Reference(Some(fn_ref)) = fn_slot else {
+        return Ok(Some(Slot::Reference(None)));
+    };
+    let fn_class = heap.get(fn_ref)?.class_name.clone();
+    // Find old value
+    let fields = &heap.get(this_ref)?.fields;
+    let old_value = hashmap_find_key(fields, key, heap)
+        .and_then(|ki| fields.get(ki + 1).copied())
+        .unwrap_or(Slot::Reference(None));
+    // Call BiFunction.apply(key, oldValue)
+    let new_value = ops.invoke(
+        heap,
+        out,
+        &fn_class,
+        "apply",
+        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
+        vec![fn_slot, key, old_value],
+    )?;
+    // null return means remove the key
+    let is_null = matches!(new_value, None | Some(Slot::Reference(None)));
+    let new_val = new_value.unwrap_or(Slot::Reference(None));
+    let fields2 = &heap.get(this_ref)?.fields;
+    if let Some(ki) = hashmap_find_key(fields2, key, heap) {
+        if is_null {
+            // Remove the key-value pair (swap-remove style)
+            let fields3 = &mut heap.get_mut(this_ref)?.fields;
+            fields3.remove(ki + 1);
+            fields3.remove(ki);
+            let size = match fields3.first() {
+                Some(Slot::Int(n)) => *n,
+                _ => 0,
+            };
+            if let Some(first) = fields3.first_mut() {
+                *first = Slot::Int(size - 1);
+            }
+        } else {
+            heap.get_mut(this_ref)?.fields[ki + 1] = new_val;
+        }
+    } else if !is_null {
+        let size = match heap.get(this_ref)?.fields.first().copied() {
+            Some(Slot::Int(n)) => usize::try_from(n.max(0)).unwrap_or(0),
+            _ => 0,
+        };
+        heap.get_mut(this_ref)?.fields.push(key);
+        heap.get_mut(this_ref)?.fields.push(new_val);
+        heap.get_mut(this_ref)?.fields[0] = Slot::Int(i32::try_from(size + 1).unwrap_or(i32::MAX));
+    }
+    Ok(Some(new_val))
+}
+/// Native: `HashMap.merge(K, V, BiFunction)V` — put V if absent, else merge with `BiFunction`.
+pub(crate) fn native_hashmap_merge(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let new_val_slot = extract_slot_arg(args, 2);
+    let fn_slot = extract_slot_arg(args, 3);
+    let fields = &heap.get(this_ref)?.fields;
+    let old_ki = hashmap_find_key(fields, key, heap);
+    if let Some(ki) = old_ki {
+        let old_value = fields.get(ki + 1).copied().unwrap_or(Slot::Reference(None));
+        let Slot::Reference(Some(fn_ref)) = fn_slot else {
+            return Ok(Some(old_value));
+        };
+        let fn_class = heap.get(fn_ref)?.class_name.clone();
+        let merged = ops.invoke(
+            heap,
+            out,
+            &fn_class,
+            "apply",
+            "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
+            vec![fn_slot, old_value, new_val_slot],
+        )?;
+        let merged_raw = merged.unwrap_or(Slot::Reference(None));
+        // Box primitive results so the stored value is always a Reference (matches Java generics)
+        let merged_val = box_primitive_slot(merged_raw, heap);
+        heap.get_mut(this_ref)?.fields[ki + 1] = merged_val;
+        Ok(Some(merged_val))
+    } else {
+        // Key absent — insert new value
+        let size = match heap.get(this_ref)?.fields.first().copied() {
+            Some(Slot::Int(n)) => usize::try_from(n.max(0)).unwrap_or(0),
+            _ => 0,
+        };
+        heap.get_mut(this_ref)?.fields.push(key);
+        heap.get_mut(this_ref)?.fields.push(new_val_slot);
+        heap.get_mut(this_ref)?.fields[0] = Slot::Int(i32::try_from(size + 1).unwrap_or(i32::MAX));
+        Ok(Some(new_val_slot))
+    }
+}
+/// Native: `StringJoiner.<init>(CharSequence)V` — delimiter only.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_stringjoiner_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let delim_slot = extract_slot_arg(args, 1);
+    let empty_ref = heap.allocate_string(String::new());
+    let prefix_ref = heap.allocate_string(String::new());
+    let suffix_ref = heap.allocate_string(String::new());
+    heap.get_mut(this_ref)?.fields[0] = delim_slot;
+    heap.get_mut(this_ref)?.fields[1] = Slot::Reference(Some(prefix_ref));
+    heap.get_mut(this_ref)?.fields[2] = Slot::Reference(Some(suffix_ref));
+    heap.get_mut(this_ref)?.fields[3] = Slot::Reference(Some(empty_ref));
+    Ok(None)
+}
+/// Native: `StringJoiner.<init>(CharSequence,CharSequence,CharSequence)V` — delim + prefix + suffix.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_stringjoiner_init_prefix_suffix(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let delim_slot = extract_slot_arg(args, 1);
+    let prefix_slot = extract_slot_arg(args, 2);
+    let suffix_slot = extract_slot_arg(args, 3);
+    let empty_ref = heap.allocate_string(String::new());
+    heap.get_mut(this_ref)?.fields[0] = delim_slot;
+    heap.get_mut(this_ref)?.fields[1] = prefix_slot;
+    heap.get_mut(this_ref)?.fields[2] = suffix_slot;
+    heap.get_mut(this_ref)?.fields[3] = Slot::Reference(Some(empty_ref));
+    Ok(None)
+}
+/// Native: `StringJoiner.add(CharSequence)StringJoiner` — append element.
+pub(crate) fn native_stringjoiner_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let elem = extract_slot_arg(args, 1);
+    heap.get_mut(this_ref)?.fields.push(elem);
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+/// Native: `StringJoiner.setEmptyValue(CharSequence)StringJoiner`.
+pub(crate) fn native_stringjoiner_set_empty_value(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let empty_slot = extract_slot_arg(args, 1);
+    heap.get_mut(this_ref)?.fields[3] = empty_slot;
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+/// Native: `StringJoiner.toString()String` — builds the joined result.
+///
+/// ⚡ Bolt Optimization: Eliminated intermediate `Vec<String>` allocation and format
+/// macro overhead by appending directly to a single String buffer.
+pub(crate) fn native_stringjoiner_tostring(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let fields = &heap.get(this_ref)?.fields;
+    // elements start at index 4
+    let elems: Vec<Slot> = fields.get(4..).map(<[Slot]>::to_vec).unwrap_or_default();
+    if elems.is_empty() {
+        let empty_slot = fields.get(3).copied().unwrap_or(Slot::Reference(None));
+        let prefix = slot_to_string(
+            fields.get(1).copied().unwrap_or(Slot::Reference(None)),
+            heap,
+        );
+        let suffix = slot_to_string(
+            fields.get(2).copied().unwrap_or(Slot::Reference(None)),
+            heap,
+        );
+        let empty = slot_to_string(empty_slot, heap);
+        // If prefix+suffix are both empty, return emptyValue; otherwise prefix+suffix
+        let result = if prefix.is_empty() && suffix.is_empty() {
+            empty
+        } else {
+            format!("{prefix}{suffix}")
+        };
+        let r = heap.allocate_string(result);
+        return Ok(Some(Slot::Reference(Some(r))));
+    }
+    let delim = slot_to_string(
+        fields.first().copied().unwrap_or(Slot::Reference(None)),
+        heap,
+    );
+    let prefix = slot_to_string(
+        fields.get(1).copied().unwrap_or(Slot::Reference(None)),
+        heap,
+    );
+    let suffix = slot_to_string(
+        fields.get(2).copied().unwrap_or(Slot::Reference(None)),
+        heap,
+    );
+    // ⚡ Bolt: Eliminate intermediate Vec<String> allocation and format! macro overhead
+    // by appending directly to a single String buffer.
+    let mut result = String::new();
+    result.push_str(&prefix);
+    for (i, elem) in elems.iter().enumerate() {
+        if i > 0 {
+            result.push_str(&delim);
+        }
+        result.push_str(&slot_to_string(*elem, heap));
+    }
+    result.push_str(&suffix);
+    let r = heap.allocate_string(result);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `StringJoiner.length()I` — length of the `toString()` result.
+pub(crate) fn native_stringjoiner_length(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    // Reuse toString and measure
+    let result = native_stringjoiner_tostring(args, heap, out, control)?;
+    let len = match result {
+        Some(Slot::Reference(Some(r))) => heap.get(r)?.string_value.as_deref().unwrap_or("").len(),
+        _ => 0,
+    };
+    Ok(Some(Slot::Int(i32::try_from(len).unwrap_or(i32::MAX))))
+}
+pub(crate) fn native_hashmap_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.is_empty() {
+        obj.fields.push(Slot::Int(0));
+    } else {
+        obj.fields[0] = Slot::Int(0);
+    }
+    Ok(None)
+}
+/// Native: `HashMap.put(Object, Object)Object` — inserts or updates a key-value pair.
+/// Returns the old value if the key was already present, or null if it is new.
+pub(crate) fn native_hashmap_put(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let val = extract_slot_arg(args, 2);
+    // We just find the index, without taking fields ownership yet
+    let i_opt = find_hashmap_entry_index(&heap.get(this_ref)?.fields, &key, heap);
+
+    if let Some(i) = i_opt {
+        let old = heap.get(this_ref)?.fields[i + 1];
+        heap.get_mut(this_ref)?.fields[i + 1] = val;
+        return Ok(Some(old));
+    }
+
+    // New key — append pair and bump size.
+    let obj = heap.get_mut(this_ref)?;
+    match obj.fields.first_mut() {
+        Some(Slot::Int(sz)) => *sz += 1,
+        _ => return Err(Error::NullPointerException),
+    }
+    obj.fields.push(key);
+    obj.fields.push(val);
+    Ok(Some(Slot::Reference(None)))
+}
+/// Native: `HashMap.get(Object)Object` — returns value for key, or null if absent.
+pub(crate) fn native_hashmap_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let fields = &heap.get(this_ref)?.fields;
+
+    Ok(Some(
+        find_hashmap_entry_index(fields, &key, heap)
+            .map_or(Slot::Reference(None), |i| fields[i + 1]),
+    ))
+}
+/// Native: `HashMap.containsKey(Object)Z` — returns 1 if key present, 0 otherwise.
+pub(crate) fn native_hashmap_contains_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let fields = &heap.get(this_ref)?.fields;
+
+    if find_hashmap_entry_index(fields, &key, heap).is_some() {
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+/// Native: `HashMap.size()I` — returns entry count from fields\[0\].
+pub(crate) fn native_hashmap_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => Ok(Some(Slot::Int(*n))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+/// Native: `HashMap.remove(Object)Object` — removes a key-value pair, returns old value or null.
+/// Uses swap-remove (swaps target pair with last pair) for O(1) deletion.
+pub(crate) fn native_hashmap_remove(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let i_opt = find_hashmap_entry_index(&heap.get(this_ref)?.fields, &key, heap);
+
+    if let Some(i) = i_opt {
+        let old_val = heap.get(this_ref)?.fields[i + 1];
+        let obj = heap.get_mut(this_ref)?;
+        let last_val_idx = obj.fields.len() - 1;
+        let last_key_idx = obj.fields.len() - 2;
+        obj.fields.swap(i + 1, last_val_idx);
+        obj.fields.swap(i, last_key_idx);
+        obj.fields.truncate(obj.fields.len() - 2);
+        match obj.fields.first_mut() {
+            Some(Slot::Int(sz)) => *sz -= 1,
+            _ => return Err(Error::NullPointerException),
+        }
+        Ok(Some(old_val))
+    } else {
+        Ok(Some(Slot::Reference(None)))
+    }
+}
+/// Native: `HashMap.isEmpty()Z` — returns 1 if size == 0, else 0.
+pub(crate) fn native_hashmap_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(0)) => Ok(Some(Slot::Int(1))),
+        Some(Slot::Int(_)) => Ok(Some(Slot::Int(0))),
+        _ => Ok(Some(Slot::Int(1))),
+    }
+}
+/// Native: `HashMap.getOrDefault(Object, Object)Object` — returns value for key, or default if absent.
+pub(crate) fn native_hashmap_get_or_default(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let default = extract_slot_arg(args, 2);
+    let fields = &heap.get(this_ref)?.fields;
+
+    Ok(Some(
+        find_hashmap_entry_index(fields, &key, heap).map_or(default, |i| fields[i + 1]),
+    ))
+}
+pub(crate) fn native_properties_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    init_properties_with_defaults(heap, this_ref, Slot::Reference(None))?;
+    Ok(None)
+}
+pub(crate) fn native_properties_init_defaults(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let defaults = extract_slot_arg(args, 1);
+    init_properties_with_defaults(heap, this_ref, defaults)?;
+    Ok(None)
+}
+pub(crate) fn native_properties_set_property(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key_ref = extract_ref_arg(args, 1)?;
+    let value_ref = extract_ref_arg(args, 2)?;
+    let _key = string_value_from_ref(heap, key_ref)?;
+    let _value = string_value_from_ref(heap, value_ref)?;
+    let old = properties_put_slots(
+        heap,
+        this_ref,
+        Slot::Reference(Some(key_ref)),
+        Slot::Reference(Some(value_ref)),
+    )?;
+    Ok(Some(old))
+}
+pub(crate) fn native_properties_get_property(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key_ref = extract_ref_arg(args, 1)?;
+    let _key = string_value_from_ref(heap, key_ref)?;
+    Ok(Some(properties_get_property_slot(
+        heap,
+        this_ref,
+        Slot::Reference(Some(key_ref)),
+    )?))
+}
+pub(crate) fn native_properties_get_property_default(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key_ref = extract_ref_arg(args, 1)?;
+    let default = extract_slot_arg(args, 2);
+    let _key = string_value_from_ref(heap, key_ref)?;
+    let value = properties_get_property_slot(heap, this_ref, Slot::Reference(Some(key_ref)))?;
+    if matches!(value, Slot::Reference(None)) {
+        Ok(Some(default))
+    } else {
+        Ok(Some(value))
+    }
+}
+pub(crate) fn native_properties_load(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let file_id = properties_stream_id_from_slot(extract_slot_arg(args, 1), heap)?;
+    let mut bytes = Vec::new();
+    let mut buf = [0u8; 4096];
+    loop {
+        let n = heap.read_host_file_bytes(file_id, &mut buf)?;
+        if n < 0 {
+            break;
+        }
+        let n = usize::try_from(n).unwrap_or(0);
+        bytes.extend_from_slice(&buf[..n]);
+    }
+
+    for (key, value) in parse_properties_bytes(&bytes)? {
+        properties_put_string_pair(heap, this_ref, key, value)?;
+    }
+    Ok(None)
+}
+pub(crate) fn native_properties_store(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let file_id = properties_stream_id_from_slot(extract_slot_arg(args, 1), heap)?;
+    let comment_slot = extract_slot_arg(args, 2);
+    let comment = match comment_slot {
+        Slot::Reference(Some(comment_ref)) => Some(string_value_from_ref(heap, comment_ref)?),
+        Slot::Reference(None) => None,
+        _ => return Err(Error::TypeMismatch { expected: "Reference", got: "other" }),
+    };
+    let rendered = render_properties_store(heap, this_ref, comment.as_deref())?;
+    write_iso_8859_1_ascii(heap, file_id, &rendered)?;
+    Ok(None)
+}
+pub(crate) fn native_properties_property_names(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let mut names = Vec::new();
+    properties_collect_name_slots(heap, this_ref, &mut names)?;
+    let enum_ref = heap.allocate(
+        "duke/util/PropertiesEnumeration".to_string(),
+        PROPERTIES_ENUM_NAMES_START + names.len(),
+    );
+    heap.write_field(enum_ref, PROPERTIES_ENUM_INDEX_FIELD, Slot::Int(0))?;
+    heap.write_field(
+        enum_ref,
+        PROPERTIES_ENUM_COUNT_FIELD,
+        Slot::Int(i32::try_from(names.len()).unwrap_or(i32::MAX)),
+    )?;
+    for (idx, name) in names.iter().copied().enumerate() {
+        heap.write_field(enum_ref, PROPERTIES_ENUM_NAMES_START + idx, name)?;
+    }
+    Ok(Some(Slot::Reference(Some(enum_ref))))
+}
+pub(crate) fn native_properties_string_property_names(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let mut names = Vec::new();
+    properties_collect_name_slots(heap, this_ref, &mut names)?;
+    let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(set_ref))], heap, out, control)?;
+    for name in names {
+        native_hashset_add(&[Slot::Reference(Some(set_ref)), name], heap, out, control)?;
+    }
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+pub(crate) fn native_properties_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.get(PROPERTIES_SIZE_FIELD) {
+        Some(Slot::Int(size)) => *size,
+        _ => 0,
+    };
+    Ok(Some(Slot::Int(size)))
+}
+pub(crate) fn native_properties_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let is_empty = matches!(
+        heap.get(this_ref)?.fields.get(PROPERTIES_SIZE_FIELD),
+        Some(Slot::Int(0)) | None
+    );
+    Ok(Some(Slot::Int(i32::from(is_empty))))
+}
+pub(crate) fn native_properties_contains_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let fields = &heap.get(this_ref)?.fields;
+    Ok(Some(Slot::Int(i32::from(
+        properties_entry_index(fields, &key, heap).is_some(),
+    ))))
+}
+pub(crate) fn native_properties_clear(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    ensure_properties_layout(heap, this_ref)?;
+    let props = heap.get_mut(this_ref)?;
+    props.fields[PROPERTIES_SIZE_FIELD] = Slot::Int(0);
+    props.fields.truncate(PROPERTIES_ENTRIES_START);
+    Ok(None)
+}
+pub(crate) fn native_properties_remove(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let idx_opt = {
+        let fields = &heap.get(this_ref)?.fields;
+        properties_entry_index(fields, &key, heap)
+    };
+
+    if let Some(idx) = idx_opt {
+        let old = heap.get(this_ref)?.fields[idx + 1];
+        let props = heap.get_mut(this_ref)?;
+        let last_value_idx = props.fields.len() - 1;
+        let last_key_idx = props.fields.len() - 2;
+        props.fields.swap(idx + 1, last_value_idx);
+        props.fields.swap(idx, last_key_idx);
+        props.fields.truncate(props.fields.len() - 2);
+        if let Some(Slot::Int(size)) = props.fields.get_mut(PROPERTIES_SIZE_FIELD) {
+            *size -= 1;
+        }
+        Ok(Some(old))
+    } else {
+        Ok(Some(Slot::Reference(None)))
+    }
+}
+pub(crate) fn native_properties_key_set(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(set_ref))], heap, out, control)?;
+    for (key, _) in properties_local_entries(heap, this_ref)? {
+        native_hashset_add(&[Slot::Reference(Some(set_ref)), key], heap, out, control)?;
+    }
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+pub(crate) fn native_properties_values(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
+    native_arraylist_init(&[Slot::Reference(Some(list_ref))], heap, out, control)?;
+    for (_, value) in properties_local_entries(heap, this_ref)? {
+        native_arraylist_add(&[Slot::Reference(Some(list_ref)), value], heap, out, control)?;
+    }
+    Ok(Some(Slot::Reference(Some(list_ref))))
+}
+pub(crate) fn native_properties_entry_set(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(set_ref))], heap, out, control)?;
+    for (key, value) in properties_local_entries(heap, this_ref)? {
+        let entry_ref = heap.allocate("java/util/Map$Entry".to_string(), 2);
+        {
+            let entry = heap.get_mut(entry_ref)?;
+            entry.fields[0] = key;
+            entry.fields[1] = value;
+        }
+        native_hashset_add(
+            &[
+                Slot::Reference(Some(set_ref)),
+                Slot::Reference(Some(entry_ref)),
+            ],
+            heap,
+            out,
+            control,
+        )?;
+    }
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+pub(crate) fn native_properties_to_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let mut rendered = String::from("{");
+    let mut first = true;
+    for (key, value) in properties_local_entries(heap, this_ref)? {
+        if first {
+            first = false;
+        } else {
+            rendered.push_str(", ");
+        }
+        rendered.push_str(&slot_to_string(key, heap));
+        rendered.push('=');
+        rendered.push_str(&slot_to_string(value, heap));
+    }
+    rendered.push('}');
+    let string_ref = heap.allocate_string(rendered);
+    Ok(Some(Slot::Reference(Some(string_ref))))
+}
+pub(crate) fn native_set_of(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
+    native_hashset_init(&[Slot::Reference(Some(set_ref))], heap, out, control)?;
+
+    match extract_slot_arg(args, 0) {
+        Slot::Reference(Some(array_ref)) => {
+            let elements = heap.get(array_ref)?.fields.clone();
+            for element in elements {
+                native_hashset_add(
+                    &[Slot::Reference(Some(set_ref)), element],
+                    heap,
+                    out,
+                    control,
+                )?;
+            }
+        }
+        Slot::Reference(None) => return Err(Error::NullPointerException),
+        _ => {
+            return Err(Error::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    }
+
+    Ok(Some(Slot::Reference(Some(set_ref))))
+}
+pub(crate) fn native_hashset_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.is_empty() {
+        obj.fields.push(Slot::Int(0));
+    } else {
+        obj.fields[0] = Slot::Int(0);
+    }
+    Ok(None)
+}
+pub(crate) fn native_hashset_init_from_collection(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let mut this_ref = extract_ref_arg(args, 0)?;
+    let collection_ref = extract_ref_arg(args, 1)?;
+    native_hashset_init(&[Slot::Reference(Some(this_ref))], heap, output, control)?;
+
+    let collection_class = heap.get(collection_ref)?.class_name.clone();
+    let elements: Vec<Slot> = if collection_class == "java/util/HashSet" {
+        heap.get(collection_ref)?
+            .fields
+            .iter()
+            .skip(1)
+            .copied()
+            .collect()
+    } else {
+        let array_slot = ops.invoke(
+            heap,
+            output,
+            &collection_class,
+            "toArray",
+            "()[Ljava/lang/Object;",
+            vec![Slot::Reference(Some(collection_ref))],
+        )?;
+        let Some(Slot::Reference(Some(array_ref))) = array_slot else {
+            return Err(Error::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        };
+        patch_forwarded_ref_if_needed(heap, &mut this_ref);
+        heap.get(array_ref)?.fields.clone()
+    };
+
+    for element in elements {
+        native_hashset_add(
+            &[Slot::Reference(Some(this_ref)), element],
+            heap,
+            output,
+            control,
+        )?;
+    }
+    Ok(None)
+}
+/// Native: `HashSet.add(Object)Z` — adds element if not already present.
+/// Returns 1 if added, 0 if element was already in the set.
+pub(crate) fn native_hashset_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let element = extract_slot_arg(args, 1);
+    // fields[0] = size, fields[1..] = elements
+    if find_hashset_entry_index(&heap.get(this_ref)?.fields, &element, heap).is_some() {
+        return Ok(Some(Slot::Int(0))); // duplicate
+    }
+
+    let obj = heap.get_mut(this_ref)?;
+    match obj.fields.first_mut() {
+        Some(Slot::Int(sz)) => *sz += 1,
+        _ => return Err(Error::NullPointerException),
+    }
+    obj.fields.push(element);
+    Ok(Some(Slot::Int(1)))
+}
+/// Native: `HashSet.contains(Object)Z` — returns 1 if element is present, 0 otherwise.
+pub(crate) fn native_hashset_contains(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let element = extract_slot_arg(args, 1);
+    let fields = &heap.get(this_ref)?.fields;
+
+    if find_hashset_entry_index(fields, &element, heap).is_some() {
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+/// Native: `HashSet.remove(Object)Z` — removes element if present, returns 1 if removed, 0 if absent.
+/// Uses swap-remove (swaps target with last element) for O(1) deletion.
+pub(crate) fn native_hashset_remove(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let element = extract_slot_arg(args, 1);
+    let i_opt = find_hashset_entry_index(&heap.get(this_ref)?.fields, &element, heap);
+
+    if let Some(i) = i_opt {
+        let obj = heap.get_mut(this_ref)?;
+        let last_idx = obj.fields.len() - 1;
+        obj.fields.swap(i, last_idx);
+        obj.fields.truncate(obj.fields.len() - 1);
+        match obj.fields.first_mut() {
+            Some(Slot::Int(sz)) => *sz -= 1,
+            _ => return Err(Error::NullPointerException),
+        }
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+/// Native: `HashSet.size()I` — returns element count from fields\[0\].
+pub(crate) fn native_hashset_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => Ok(Some(Slot::Int(*n))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+/// Native: `HashSet.isEmpty()Z` — returns 1 if size == 0, else 0.
+pub(crate) fn native_hashset_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(0)) => Ok(Some(Slot::Int(1))),
+        Some(Slot::Int(_)) => Ok(Some(Slot::Int(0))),
+        _ => Ok(Some(Slot::Int(1))),
+    }
+}
+/// Native: `HashSet.iterator()Iterator` — creates a `HashSetIterator`.
+pub(crate) fn native_hashset_iterator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let iter_ref = heap.allocate("duke/util/HashSetIterator".to_string(), 2);
+    {
+        let iter_obj = heap.get_mut(iter_ref)?;
+        iter_obj.fields[0] = Slot::Reference(Some(this_ref));
+        iter_obj.fields[1] = Slot::Int(0);
+    }
+    Ok(Some(Slot::Reference(Some(iter_ref))))
+}
+/// Native: `HashSet.stream()Stream` — wraps elements into a `duke/util/Stream`.
+pub(crate) fn native_hashset_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(this_ref)?.fields[1..=size].to_vec();
+    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
+    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(i32::try_from(size).unwrap_or(0));
+    heap.get_mut(stream_ref)?.fields.extend(elems);
+    Ok(Some(Slot::Reference(Some(stream_ref))))
+}
+/// Native: `LinkedList.stream()Stream` — wraps elements into a `duke/util/Stream`.
+pub(crate) fn native_linked_list_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_stream(args, heap, out, control)
+}
+/// Native: `Collections.nCopies(int, Object)List` — returns a list of N copies of an element.
+pub(crate) fn native_collections_n_copies(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let n = usize::try_from(extract_int_arg(args, 0)?.max(0)).unwrap_or(0);
+    let elem = extract_slot_arg(args, 1);
+    let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
+    heap.get_mut(list_ref)?.fields[0] = Slot::Int(i32::try_from(n).unwrap_or(0));
+    for _ in 0..n {
+        heap.get_mut(list_ref)?.fields.push(elem);
+    }
+    Ok(Some(Slot::Reference(Some(list_ref))))
+}
+/// Native: `ArrayList.forEach(Consumer)V` — invokes consumer.accept(elem) for each element.
+pub(crate) fn native_arraylist_for_each(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let Slot::Reference(Some(consumer_ref)) = extract_slot_arg(args, 1)
+    else {
+        return Ok(None);
+    };
+    let size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(list_ref)?.fields[1..=size].to_vec();
+    let consumer_class = heap.get(consumer_ref)?.class_name.clone();
+    for elem in elems {
+        ops.invoke(
+            heap,
+            out,
+            &consumer_class,
+            "accept",
+            "(Ljava/lang/Object;)V",
+            vec![Slot::Reference(Some(consumer_ref)), elem],
+        )?;
+    }
+    Ok(None)
+}
+/// Native: `Arrays.toString(int[])String` — formats as `[1, 2, 3]`.
+pub(crate) fn native_arrays_to_string_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    use std::fmt::Write as FmtWrite;
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let fields = heap.get(arr_ref)?.fields.clone();
+
+    // ⚡ Bolt: Eliminate intermediate Vec<String> allocation, format! macro overhead,
+    // and .join() by appending directly to a single String buffer.
+    let mut result = String::with_capacity(fields.len() * 4 + 2);
+    result.push('[');
+    for (i, s) in fields.iter().enumerate() {
+        if i > 0 {
+            result.push_str(", ");
+        }
+        match s {
+            Slot::Int(n) => { let _ = write!(result, "{n}"); },
+            _ => result.push('0'),
+        }
+    }
+    result.push(']');
+
+    let r = heap.allocate_string(result);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Arrays.toString(Object[])String` — formats as `[a, b, c]`.
+pub(crate) fn native_arrays_to_string_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    use std::fmt::Write as FmtWrite;
+    let arr_ref = extract_ref_arg(args, 0)?;
+    let fields = heap.get(arr_ref)?.fields.clone();
+
+    // ⚡ Bolt: Eliminate intermediate Vec<String> allocation, format! macro overhead,
+    // and .join() by appending directly to a single String buffer.
+    let mut result = String::with_capacity(fields.len() * 8 + 2);
+    result.push('[');
+    for (i, s) in fields.iter().enumerate() {
+        if i > 0 {
+            result.push_str(", ");
+        }
+        match s {
+            Slot::Reference(Some(r)) => {
+                let part = heap
+                    .get(*r)
+                    .ok()
+                    .and_then(|o| o.string_value.clone())
+                    .unwrap_or_else(|| "null".to_string());
+                result.push_str(&part);
+            }
+            Slot::Reference(None) => result.push_str("null"),
+            Slot::Int(n) => {
+                let _ = write!(result, "{n}");
+            }
+            _ => result.push('?'),
+        }
+    }
+    result.push(']');
+
+    let r = heap.allocate_string(result);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `HashMap.replace(Object, Object)Object` — updates value for existing key,
+/// returns the old value or null if key was absent.
+pub(crate) fn native_hashmap_replace(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let new_val = extract_slot_arg(args, 2);
+    let i_opt = find_hashmap_entry_index(&heap.get(this_ref)?.fields, &key, heap);
+    if let Some(i) = i_opt {
+        let old = heap.get(this_ref)?.fields[i + 1];
+        heap.get_mut(this_ref)?.fields[i + 1] = new_val;
+        Ok(Some(old))
+    } else {
+        Ok(Some(Slot::Reference(None)))
+    }
+}
+/// Native: `Collections.swap(List, int, int)V` — swaps elements at indices i and j.
+pub(crate) fn native_collections_swap(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let list_ref = extract_ref_arg(args, 0)?;
+    let i = match args.get(1) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => return Ok(None),
+    };
+    let j = match args.get(2) {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => return Ok(None),
+    };
+    // fields[0] = size, elements at fields[1..=size]
+    let fi = i + 1;
+    let fj = j + 1;
+    let fields = heap.get(list_ref)?.fields.clone();
+    let len = fields.len();
+    if fi < len && fj < len {
+        let vi = fields[fi];
+        let vj = fields[fj];
+        heap.get_mut(list_ref)?.fields[fi] = vj;
+        heap.get_mut(list_ref)?.fields[fj] = vi;
+    }
+    Ok(None)
+}
+/// Native: `Collections.unmodifiableMap(Map)Map` — identity stub (we have no mutation checks).
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_collections_unmodifiable_map(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let src_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Ok(Some(Slot::Reference(None))),
+    };
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let src_class = heap.get(src_ref)?.class_name.clone();
+    // Copy data into an UnmodifiableMap wrapper (same layout as HashMap)
+    let r = heap.allocate("java/util/UnmodifiableMap".to_string(), src_fields.len());
+    let r_fields = &mut heap.get_mut(r)?.fields;
+    *r_fields = src_fields;
+    drop(src_class);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Comparator.comparingLong(ToLongFunction)Comparator` — wraps key extractor.
+/// Creates a `duke/util/ComparingLongComparator` with `fields[0] = fn_ref`.
+pub(crate) fn native_comparator_comparing_long(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let fn_ref = extract_ref_arg(args, 0)?;
+    let r = heap.allocate("duke/util/ComparingLongComparator".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = Slot::Reference(Some(fn_ref));
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Creates a `duke/util/ComparingDoubleComparator` with `fields[0] = fn_ref`.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_comparator_comparing_double(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let fn_ref = extract_ref_arg(args, 0)?;
+    let r = heap.allocate("duke/util/ComparingDoubleComparator".to_string(), 1);
+    heap.get_mut(r)?.fields[0] = Slot::Reference(Some(fn_ref));
+    Ok(Some(Slot::Reference(Some(r))))
+}
+/// Native: `Optional.or(Supplier<Optional>)Optional` (Java 9) —
+/// returns this Optional if present; otherwise invokes supplier and returns its result.
+pub(crate) fn native_optional_or(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let opt_ref = extract_ref_arg(args, 0)?;
+    let supplier_slot = extract_slot_arg(args, 1);
+    let value = extract_first_field_arg(heap, opt_ref)?;
+    // If present (non-null value stored), return self
+    if !matches!(value, Slot::Reference(None)) {
+        return Ok(Some(Slot::Reference(Some(opt_ref))));
+    }
+    let Slot::Reference(Some(supplier_ref)) = supplier_slot else {
+        return Ok(Some(Slot::Reference(Some(opt_ref))));
+    };
+    let supplier_class = heap.get(supplier_ref)?.class_name.clone();
+    let result = ops.invoke(
+        heap,
+        out,
+        &supplier_class,
+        "get",
+        "()Ljava/lang/Object;",
+        vec![supplier_slot],
+    )?;
+    Ok(Some(result.unwrap_or(Slot::Reference(None))))
+}
+/// Native: `Optional.ifPresentOrElse(Consumer, Runnable)V` (Java 9) —
+/// if value present invokes consumer, otherwise invokes runnable.
+pub(crate) fn native_optional_if_present_or_else(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let opt_ref = extract_ref_arg(args, 0)?;
+    let consumer_slot = extract_slot_arg(args, 1);
+    let runnable_slot = extract_slot_arg(args, 2);
+    let value = extract_first_field_arg(heap, opt_ref)?;
+    if matches!(value, Slot::Reference(None)) {
+        let Slot::Reference(Some(runnable_ref)) = runnable_slot else {
+            return Ok(None);
+        };
+        let runnable_class = heap.get(runnable_ref)?.class_name.clone();
+        ops.invoke(
+            heap,
+            out,
+            &runnable_class,
+            "run",
+            "()V",
+            vec![runnable_slot],
+        )?;
+    } else {
+        let Slot::Reference(Some(consumer_ref)) = consumer_slot else {
+            return Ok(None);
+        };
+        let consumer_class = heap.get(consumer_ref)?.class_name.clone();
+        ops.invoke(
+            heap,
+            out,
+            &consumer_class,
+            "accept",
+            "(Ljava/lang/Object;)V",
+            vec![consumer_slot, value],
+        )?;
+    }
+    Ok(None)
+}
+/// Native: `Optional.stream()Stream` — returns a stream of 0 or 1 elements.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_optional_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let value = extract_first_field_arg(heap, this_ref)?;
+    let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
+    match value {
+        Slot::Reference(None) => {
+            heap.get_mut(out_ref)?.fields[0] = Slot::Int(0);
+        }
+        v => {
+            heap.get_mut(out_ref)?.fields[0] = Slot::Int(1);
+            heap.get_mut(out_ref)?.fields.push(v);
+        }
+    }
+    Ok(Some(Slot::Reference(Some(out_ref))))
+}
+/// Native: `ArrayDeque.addFirst(Object)V` — inserts element at front.
+pub(crate) fn native_arraydeque_add_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraydeque_push(args, heap, out, control)
+}
+/// Native: `ArrayDeque.addLast(Object)V` — appends element at back.
+pub(crate) fn native_arraydeque_add_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)?;
+    Ok(None)
+}
+/// Native: `ArrayDeque.offerFirst(Object)Z` — inserts at front, returns true.
+pub(crate) fn native_arraydeque_offer_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraydeque_push(args, heap, out, control)?;
+    Ok(Some(Slot::Int(1)))
+}
+/// Native: `ArrayDeque.offerLast(Object)Z` — appends at back, returns true.
+pub(crate) fn native_arraydeque_offer_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraylist_add(args, heap, out, control)?;
+    Ok(Some(Slot::Int(1)))
+}
+/// Native: `ArrayDeque.peekFirst()Object` — same as peek (front element, null if empty).
+pub(crate) fn native_arraydeque_peek_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraydeque_peek(args, heap, out, control)
+}
+/// Native: `ArrayDeque.peekLast()Object` — returns last element without removing; null if empty.
+pub(crate) fn native_arraydeque_peek_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    Ok(Some(heap.get(this_ref)?.fields[size]))
+}
+/// Native: `ArrayDeque.pollFirst()Object` — same as poll (remove front, null if empty).
+pub(crate) fn native_arraydeque_poll_first(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    native_arraydeque_poll(args, heap, out, control)
+}
+/// Native: `ArrayDeque.pollLast()Object` — removes and returns last element; null if empty.
+pub(crate) fn native_arraydeque_poll_last(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    if size == 0 {
+        return Ok(Some(Slot::Reference(None)));
+    }
+    let elem = heap.get_mut(this_ref)?.fields.remove(size);
+    let new_size = i32::try_from(size - 1).unwrap_or(0);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(new_size);
+    Ok(Some(elem))
+}
+/// Native: `ArrayDeque.contains(Object)Z` — returns true if element is present.
+pub(crate) fn native_arraydeque_contains(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let target = extract_slot_arg(args, 1);
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(this_ref)?.fields[1..=size].to_vec();
+    let found = elems.iter().any(|e| slots_equal(e, &target, heap));
+    Ok(Some(Slot::Int(i32::from(found))))
+}
+/// Native: `ArrayDeque.stream()Stream` — returns elements as an eager Stream.
+pub(crate) fn native_arraydeque_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(this_ref)?.fields[1..=size].to_vec();
+    let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
+    heap.get_mut(out_ref)?.fields[0] = Slot::Int(i32::try_from(size).unwrap_or(0));
+    heap.get_mut(out_ref)?.fields.extend(elems);
+    Ok(Some(Slot::Reference(Some(out_ref))))
+}
+/// Native: `ArrayDeque.forEach(Consumer)V` — invokes consumer for each element.
+pub(crate) fn native_arraydeque_for_each(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let consumer_slot = extract_slot_arg(args, 1);
+    let Slot::Reference(Some(cons_ref)) = consumer_slot else {
+        return Err(Error::NullPointerException);
+    };
+    let cons_class = heap.get(cons_ref)?.class_name.clone();
+    let size = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
+        _ => 0,
+    };
+    let elems: Vec<Slot> = heap.get(this_ref)?.fields[1..=size].to_vec();
+    for elem in elems {
+        ops.invoke(
+            heap,
+            out,
+            &cons_class,
+            "accept",
+            "(Ljava/lang/Object;)V",
+            vec![consumer_slot, elem],
+        )?;
+    }
+    Ok(None)
+}
+/// Native: `ArrayDeque.clear()V` — removes all elements.
+pub(crate) fn native_arraydeque_clear(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    heap.get_mut(this_ref)?.fields.truncate(1);
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+/// Native: `Collections.disjoint(Collection, Collection) -> boolean`
+/// Returns true if the two collections have no elements in common.
+pub(crate) fn native_collections_disjoint(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a_ref = extract_ref_arg(args, 0)?;
+    let b_ref = extract_ref_arg(args, 1)?;
+    // Both use ArrayList/HashSet layout: fields[0]=size, fields[1..=size]=elements
+    let a_size = match heap.get(a_ref)?.fields.first() {
+        Some(Slot::Int(v)) => usize::try_from(*v).unwrap_or(0),
+        _ => 0,
+    };
+    let b_size = match heap.get(b_ref)?.fields.first() {
+        Some(Slot::Int(v)) => usize::try_from(*v).unwrap_or(0),
+        _ => 0,
+    };
+    let a_max = a_size.min(heap.get(a_ref)?.fields.len().saturating_sub(1));
+    let a_elems: Vec<Slot> = if a_max >= 1 {
+        heap.get(a_ref)?.fields[1..=a_max].to_vec()
+    } else {
+        Vec::new()
+    };
+    let b_max = b_size.min(heap.get(b_ref)?.fields.len().saturating_sub(1));
+    let b_elems: Vec<Slot> = if b_max >= 1 {
+        heap.get(b_ref)?.fields[1..=b_max].to_vec()
+    } else {
+        Vec::new()
+    };
+    let disjoint = a_elems
+        .iter()
+        .all(|a| !b_elems.iter().any(|b| slots_equal(a, b, heap)));
+    Ok(Some(Slot::Int(i32::from(disjoint))))
+}
+/// Native: `HashMap.computeIfPresent(K, BiFunction<K,V,V>) -> V`
+/// If key is present, applies the function to (key, `old_value`); replaces with result.
+/// If function returns null, removes the key.
+pub(crate) fn native_hashmap_compute_if_present(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+    ops: &mut dyn CallbackOps,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    // Look up existing value.
+    let existing = native_hashmap_get(&[Slot::Reference(Some(this_ref)), key], heap, out, control)?;
+    let old_value = match existing {
+        Some(v) if !matches!(v, Slot::Reference(None)) => v,
+        _ => return Ok(Some(Slot::Reference(None))),
+    };
+    // Key present — invoke the remapping function.
+    let fn_ref = extract_ref_arg(args, 2)?;
+    let fn_class = heap.get(fn_ref)?.class_name.clone();
+    let new_value = ops.invoke(
+        heap,
+        out,
+        &fn_class,
+        "apply",
+        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
+        vec![Slot::Reference(Some(fn_ref)), key, old_value],
+    )?;
+    match new_value {
+        Some(v) if !matches!(v, Slot::Reference(None)) => {
+            native_hashmap_put(
+                &[Slot::Reference(Some(this_ref)), key, v],
+                heap,
+                out,
+                control,
+            )?;
+            Ok(Some(v))
+        }
+        _ => {
+            // null return → remove key
+            native_hashmap_remove(&[Slot::Reference(Some(this_ref)), key], heap, out, control)?;
+            Ok(Some(Slot::Reference(None)))
+        }
+    }
+}
+/// Native: `HashMap.remove(Object, Object) -> boolean`
+/// Conditional remove: only removes if key is present AND value equals the provided value.
+pub(crate) fn native_hashmap_remove_key_value(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let this_ref = extract_ref_arg(args, 0)?;
+    let key = extract_slot_arg(args, 1);
+    let expected_val = extract_slot_arg(args, 2);
+    let fields = heap.get(this_ref)?.fields.clone();
+    if let Some(i) = find_hashmap_entry_index(&fields, &key, heap) {
+        let actual_val = fields[i + 1];
+        if slots_equal(&actual_val, &expected_val, heap) {
+            native_hashmap_remove(&[Slot::Reference(Some(this_ref)), key], heap, out, control)?;
+            return Ok(Some(Slot::Int(1)));
+        }
+    }
+    Ok(Some(Slot::Int(0)))
+}
